Re-organize legacy engine code
