+++ temba/flows/models.py
@@ -23,7 +23,7 @@ from django.contrib.auth.models import Group, User
 from django.core.cache import cache
 from django.core.files.temp import NamedTemporaryFile
 from django.db import connection as db_connection, models, transaction
-from django.db.models import Max, Q, QuerySet, Sum
+from django.db.models import Max, Q, Sum
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 
@@ -33,7 +33,7 @@ from temba.assets.models import register_asset_store
 from temba.channels.models import Channel, ChannelConnection
 from temba.contacts.models import NEW_CONTACT_VARIABLE, URN, Contact, ContactField, ContactGroup, ContactURN
 from temba.locations.models import AdminBoundary
-from temba.msgs.models import DELIVERED, FAILED, FLOW, INBOX, INCOMING, OUTGOING, PENDING, Broadcast, Label, Msg
+from temba.msgs.models import FLOW, INBOX, INCOMING, Broadcast, Label, Msg
 from temba.orgs.models import Language, Org
 from temba.utils import analytics, chunk_list, json, on_transaction_commit
 from temba.utils.dates import str_to_datetime
@@ -586,144 +586,6 @@ class Flow(TembaModel):
         return name
 
     @classmethod
-    def find_and_handle(
-        cls,
-        msg,
-        started_flows=None,
-        triggered_start=False,
-        resume_parent_run=False,
-        user_input=True,
-        trigger_send=True,
-        continue_parent=True,
-    ):
-
-        if started_flows is None:
-            started_flows = []
-
-        for run in FlowRun.get_active_for_contact(msg.contact):
-            flow = run.flow
-            flow.ensure_current_version()
-
-            # it's possible Flow.start is in the process of creating a run for this contact, in which case
-            # record this message has handled so it doesn't start any new flows
-            if not run.path:  # pragma: no cover
-                if run.created_on > timezone.now() - timedelta(minutes=10):
-                    return True, []
-                else:
-                    return False, []
-
-            last_step = run.path[-1]
-            destination = Flow.get_node(flow, last_step[FlowRun.PATH_NODE_UUID], Flow.NODE_TYPE_RULESET)
-
-            # this node doesn't exist anymore, mark it as left so they leave the flow
-            if not destination:  # pragma: no cover
-                run.set_completed(exit_uuid=None)
-                return True, []
-
-            (handled, msgs) = Flow.handle_destination(
-                destination,
-                run,
-                msg,
-                started_flows,
-                user_input=user_input,
-                triggered_start=triggered_start,
-                resume_parent_run=resume_parent_run,
-                trigger_send=trigger_send,
-                continue_parent=continue_parent,
-            )
-
-            if handled:
-                analytics.gauge("temba.run_resumes")
-                return True, msgs
-
-        return False, []
-
-    @classmethod
-    def handle_destination(
-        cls,
-        destination,
-        run,
-        msg,
-        started_flows=None,
-        user_input=False,
-        triggered_start=False,
-        trigger_send=True,
-        resume_parent_run=False,
-        continue_parent=True,
-    ):
-        def add_to_path(path, uuid):
-            if uuid in path:
-                path.append(uuid)
-                raise FlowException("Flow cycle detected at runtime: %s" % path)
-            path.append(uuid)
-
-        start_time = time.time()
-        path = []
-        msgs = []
-
-        # lookup our next destination
-        handled = False
-
-        while destination:
-            result = {"handled": False}
-
-            if destination.get_step_type() == Flow.NODE_TYPE_RULESET:
-                should_pause = False
-
-                # check if we need to stop
-                if destination.is_pause():
-                    should_pause = True
-
-                if user_input or not should_pause:
-                    result = Flow.handle_ruleset(destination, run, msg, started_flows, resume_parent_run)
-                    add_to_path(path, destination.uuid)
-
-                    # add any messages generated by this ruleset
-                    msgs += result.get("msgs", [])
-
-                # if we used this input, then mark our user input as used
-                if should_pause:
-                    user_input = False
-
-                    # once we handle user input, reset our path
-                    path = []
-
-            elif destination.get_step_type() == Flow.NODE_TYPE_ACTIONSET:
-                result = Flow.handle_actionset(destination, run, msg, started_flows)
-                add_to_path(path, destination.uuid)
-
-                # add any generated messages to be sent at once
-                msgs += result.get("msgs", [])
-
-            # if this is a triggered start, we only consider user input on the first step, so clear it now
-            if triggered_start:
-                user_input = False
-
-            # lookup our next destination
-            destination = result.get("destination", None)
-
-            # if any one of our destinations handled us, consider it handled
-            if result.get("handled", False):
-                handled = True
-
-            resume_parent_run = False
-
-        # if we have a parent to continue, do so
-        if getattr(run, "continue_parent", False) and continue_parent:
-            msgs += FlowRun.continue_parent_flow_run(run, trigger_send=False, continue_parent=True)
-
-        if handled:
-            analytics.gauge("temba.flow_execution", time.time() - start_time)
-
-        # send any messages generated
-        if msgs and trigger_send:
-            msgs.sort(key=lambda message: message.created_on)
-            Msg.objects.filter(id__in=[m.id for m in msgs]).exclude(status=DELIVERED).update(status=PENDING)
-            run.flow.org.trigger_send(msgs)
-
-        return handled, msgs
-
-    @classmethod
     def handle_actionset(cls, actionset, run, msg, started_flows):
 
         # not found, escape out, but we still handled this message, user is now out of the flow
@@ -738,7 +600,7 @@ class Flow(TembaModel):
         # and onto the destination
         destination = Flow.get_node(actionset.flow, actionset.destination, actionset.destination_type)
         if destination:
-            run.flow.add_step(run, destination, exit_uuid=actionset.exit_uuid)
+            legacy.add_step(run, destination, exit_uuid=actionset.exit_uuid)
         else:
             run.set_completed(exit_uuid=actionset.exit_uuid)
 
@@ -765,7 +627,8 @@ class Flow(TembaModel):
                     run.update_expiration(timezone.now())
 
                 if flow:
-                    child_runs = flow.start(
+                    child_runs = legacy.flow_start(
+                        flow,
                         [],
                         [run.contact],
                         started_flows=started_flows,
@@ -822,7 +685,7 @@ class Flow(TembaModel):
         # Create the step for our destination
         destination = Flow.get_node(flow, result_rule.destination, result_rule.destination_type)
         if destination:
-            flow.add_step(run, destination, exit_uuid=result_rule.uuid)
+            legacy.add_step(run, destination, exit_uuid=result_rule.uuid)
 
         return dict(handled=True, destination=destination, msgs=msgs_out)
 
@@ -1408,301 +1271,6 @@ class Flow(TembaModel):
         flow_start.groups.add(*group_ids)
 
         on_transaction_commit(lambda: flow_start.async_start())
-
-    def start(
-        self,
-        groups,
-        contacts,
-        restart_participants=False,
-        started_flows=None,
-        start_msg=None,
-        extra=None,
-        flow_start=None,
-        parent_run=None,
-        interrupt=True,
-        connection=None,
-        include_active=True,
-        campaign_event=None,
-    ):
-        """
-        Starts a flow for the passed in groups and contacts.
-        """
-
-        if not getattr(settings, "USES_LEGACY_ENGINE", False):
-            raise ValueError("Use of legacy engine when USES_LEGACY_ENGINE not set")  # pragma: no cover
-
-        # build up querysets of our groups for memory efficiency
-        if isinstance(groups, QuerySet):  # pragma: no cover
-            group_qs = groups
-        else:
-            group_qs = ContactGroup.all_groups.filter(id__in=[g.id for g in groups])
-
-        # build up querysets of our contacts for memory efficiency
-        if isinstance(contacts, QuerySet):  # pragma: no cover
-            contact_qs = contacts
-        else:
-            contact_qs = Contact.objects.filter(id__in=[c.id for c in contacts])
-
-        self.ensure_current_version()
-
-        if started_flows is None:
-            started_flows = []
-
-        # prevents infinite loops
-        if self.pk in started_flows:  # pragma: needs cover
-            return []
-
-        # add this flow to our list of started flows
-        started_flows.append(self.pk)
-
-        if not self.entry_uuid:  # pragma: needs cover
-            return []
-
-        if start_msg and start_msg.id:
-            start_msg.msg_type = FLOW
-            start_msg.save(update_fields=["msg_type"])
-
-        all_contact_ids = list(
-            Contact.objects.filter(Q(all_groups__in=group_qs) | Q(id__in=contact_qs))
-            .order_by("id")
-            .values_list("id", flat=True)
-            .distinct("id")
-        )
-
-        if not restart_participants:
-            # exclude anybody who has already participated in the flow
-            already_started = set(self.runs.all().values_list("contact_id", flat=True))
-            all_contact_ids = [contact_id for contact_id in all_contact_ids if contact_id not in already_started]
-
-        if not include_active:
-            # exclude anybody who has an active flow run
-            already_active = set(
-                FlowRun.objects.filter(is_active=True, org=self.org).values_list("contact_id", flat=True)
-            )
-            all_contact_ids = [contact_id for contact_id in all_contact_ids if contact_id not in already_active]
-
-        # if we have a parent run, find any parents/grandparents that are active, we'll keep these active
-        ancestor_ids = []
-        ancestor = parent_run
-        while ancestor:
-            # we don't consider it an ancestor if it's not current in our start list
-            if ancestor.contact.id not in all_contact_ids:
-                break
-            ancestor_ids.append(ancestor.id)
-            ancestor = ancestor.parent
-
-        # for the contacts that will be started, exit any existing flow runs except system flow runs
-        for contact_batch in chunk_list(all_contact_ids, 1000):
-            active_runs = (
-                FlowRun.objects.filter(is_active=True, contact__pk__in=contact_batch)
-                .exclude(id__in=ancestor_ids)
-                .exclude(flow__is_system=True)
-            )
-            FlowRun.bulk_exit(active_runs, FlowRun.EXIT_TYPE_INTERRUPTED)
-
-        # if we are interrupting parent flow runs, mark them as completed
-        if ancestor_ids and interrupt:
-            ancestor_runs = FlowRun.objects.filter(id__in=ancestor_ids)
-            FlowRun.bulk_exit(ancestor_runs, FlowRun.EXIT_TYPE_COMPLETED)
-
-        if not all_contact_ids:
-            return []
-
-        if self.flow_type == Flow.TYPE_VOICE:  # pragma: no cover
-            raise ValueError("IVR flow '%s' no longer supported" % self.name)
-
-        return self._start_msg_flow(
-            all_contact_ids,
-            started_flows=started_flows,
-            start_msg=start_msg,
-            extra=extra,
-            flow_start=flow_start,
-            parent_run=parent_run,
-        )
-
-    def _start_msg_flow(
-        self, contact_ids, started_flows=None, start_msg=None, extra=None, flow_start=None, parent_run=None
-    ):
-        if parent_run:
-            parent_context = parent_run.build_expressions_context(contact_context=str(parent_run.contact.uuid))
-        else:
-            parent_context = None
-
-        contacts = Contact.objects.filter(id__in=contact_ids)
-        Contact.bulk_cache_initialize(self.org, contacts)
-        contact_map = {c.id: c for c in contacts}
-
-        # these fields are the initial state for our flow run
-        run_fields = {}  # this should be the default value of the FlowRun.fields
-        if extra:
-            # we keep more values in @extra for new flow runs because we might be passing the state
-            (normalized_fields, count) = FlowRun.normalize_fields(extra, settings.FLOWRUN_FIELDS_SIZE * 4)
-            run_fields = normalized_fields
-
-        # create all our flow runs for this set of contacts at once
-        batch = []
-        now = timezone.now()
-
-        for contact_id in contact_ids:
-            contact = contact_map[contact_id]
-            run = FlowRun.create(
-                self,
-                contact,
-                fields=run_fields,
-                start=flow_start,
-                created_on=now,
-                parent=parent_run,
-                parent_context=parent_context,
-                db_insert=False,
-                responded=start_msg is not None,
-            )
-            batch.append(run)
-
-        runs = FlowRun.objects.bulk_create(batch)
-
-        # build a map of contact to flow run
-        run_map = dict()
-        for run in runs:
-            run.flow = self
-            run.org = self.org
-
-            run_map[run.contact_id] = run
-
-        # update our expiration date on our runs, we do this by calculating it on one run then updating all others
-        run.update_expiration(timezone.now())
-
-        # if we have more than one run, update the others to the same expiration
-        if len(run_map) > 1:
-            FlowRun.objects.filter(id__in=[r.id for r in runs]).update(
-                expires_on=run.expires_on, modified_on=timezone.now()
-            )
-
-        # now execute our actual flow steps
-        (entry_actions, entry_rules) = (None, None)
-        if self.entry_type == Flow.NODE_TYPE_ACTIONSET:
-            entry_actions = ActionSet.objects.filter(uuid=self.entry_uuid).first()
-            if entry_actions:
-                entry_actions.flow = self
-
-        elif self.entry_type == Flow.NODE_TYPE_RULESET:
-            entry_rules = RuleSet.objects.filter(flow=self, uuid=self.entry_uuid).first()
-            if entry_rules:
-                entry_rules.flow = self
-
-        msgs_to_send = []
-
-        for run in runs:
-            contact = run.contact
-
-            # each contact maintains its own list of started flows
-            started_flows_by_contact = list(started_flows)
-            run_msgs = [start_msg] if start_msg else []
-            arrived_on = timezone.now()
-
-            try:
-                if entry_actions:
-                    run_msgs += entry_actions.execute_actions(run, start_msg, started_flows_by_contact)
-
-                    self.add_step(run, entry_actions, run_msgs, arrived_on=arrived_on)
-
-                    # and onto the destination
-                    if entry_actions.destination:
-                        destination = Flow.get_node(
-                            entry_actions.flow, entry_actions.destination, entry_actions.destination_type
-                        )
-
-                        self.add_step(run, destination, exit_uuid=entry_actions.exit_uuid)
-
-                        msg = Msg(org=self.org, contact=contact, text="", id=0)
-                        handled, step_msgs = Flow.handle_destination(
-                            destination, run, msg, started_flows_by_contact, trigger_send=False, continue_parent=False
-                        )
-                        run_msgs += step_msgs
-
-                    else:
-                        run.set_completed(exit_uuid=None)
-
-                elif entry_rules:
-                    self.add_step(run, entry_rules, run_msgs, arrived_on=arrived_on)
-
-                    # if we have a start message, go and handle the rule
-                    if start_msg:
-                        Flow.find_and_handle(start_msg, started_flows_by_contact, triggered_start=True)
-
-                    # if we didn't get an incoming message, see if we need to evaluate it passively
-                    elif not entry_rules.is_pause():
-                        # create an empty placeholder message
-                        msg = Msg(org=self.org, contact=contact, text="", id=0)
-                        handled, step_msgs = Flow.handle_destination(
-                            entry_rules, run, msg, started_flows_by_contact, trigger_send=False, continue_parent=False
-                        )
-                        run_msgs += step_msgs
-
-                # set the msgs that were sent by this run so that any caller can deal with them
-                run.start_msgs = [m for m in run_msgs if m.direction == OUTGOING]
-
-                # add these messages as ones that are ready to send
-                for msg in run_msgs:
-                    if msg.direction == OUTGOING:
-                        msgs_to_send.append(msg)
-
-            except Exception:
-                logger.error(
-                    "Failed starting flow %d for contact %d" % (self.id, contact.id), exc_info=1, extra={"stack": True}
-                )
-
-                # mark this flow as interrupted
-                run.set_interrupted()
-
-                # mark our messages as failed
-                Msg.objects.filter(id__in=[m.id for m in run_msgs if m.direction == OUTGOING]).update(status=FAILED)
-
-                # remove our msgs from our parent's concerns
-                run.start_msgs = []
-
-        # trigger our messages to be sent
-        if msgs_to_send and not parent_run:
-            # then send them off
-            msgs_to_send.sort(key=lambda message: (message.contact_id, message.created_on))
-            Msg.objects.filter(id__in=[m.id for m in msgs_to_send]).update(status=PENDING)
-
-            # trigger a sync
-            self.org.trigger_send(msgs_to_send)
-
-        return runs
-
-    def add_step(self, run, node, msgs=(), exit_uuid=None, arrived_on=None):
-        """
-        Adds a new step to the given run
-        """
-        if not arrived_on:
-            arrived_on = timezone.now()
-
-        # complete previous step
-        if run.path and exit_uuid:
-            run.path[-1][FlowRun.PATH_EXIT_UUID] = exit_uuid
-
-        # create new step
-        run.path.append(
-            {
-                FlowRun.PATH_STEP_UUID: str(uuid4()),
-                FlowRun.PATH_NODE_UUID: node.uuid,
-                FlowRun.PATH_ARRIVED_ON: arrived_on.isoformat(),
-            }
-        )
-
-        # trim path to ensure it can't grow indefinitely
-        if len(run.path) > FlowRun.PATH_MAX_STEPS:
-            run.path = run.path[len(run.path) - FlowRun.PATH_MAX_STEPS :]
-
-        update_fields = ["path", "current_node_uuid"]
-
-        if msgs:
-            run.add_messages(msgs, do_save=False)
-            update_fields += ["responded", "events"]
-
-        run.current_node_uuid = run.path[-1][FlowRun.PATH_NODE_UUID]
-        run.save(update_fields=update_fields)
 
     def get_dependencies(self):
         """
@@ -2904,6 +2472,7 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
 
     @classmethod
     def continue_parent_flow_run(cls, run, trigger_send=True, continue_parent=True):
+        from temba.flows import legacy
 
         # TODO: Remove this in favor of responded on session
         if run.responded and not run.parent.responded:
@@ -2948,7 +2517,7 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
         run.parent.save(update_fields=("child_context",))
 
         # finally, trigger our parent flow
-        (handled, msgs) = Flow.find_and_handle(
+        (handled, msgs) = legacy.find_and_handle(
             msg,
             user_input=False,
             started_flows=[run.flow, run.parent.flow],
@@ -3620,14 +3189,14 @@ class FlowRevision(SmartModel):
 
     @classmethod
     def migrate_export(cls, org, exported_json, same_site, version, to_version=None):
-        from temba.flows import flow_migrations
+        from temba.flows.legacy import migrations
 
         if not to_version:
             to_version = Flow.FINAL_LEGACY_VERSION
 
         for version in Flow.get_versions_after(version):
             version_slug = version.replace(".", "_")
-            migrate_fn = getattr(flow_migrations, "migrate_export_to_version_%s" % version_slug, None)
+            migrate_fn = getattr(migrations, "migrate_export_to_version_%s" % version_slug, None)
 
             if migrate_fn:
                 exported_json = migrate_fn(exported_json, org, same_site)
@@ -3641,7 +3210,7 @@ class FlowRevision(SmartModel):
                 exported_json["flows"] = flows
 
             else:
-                migrate_fn = getattr(flow_migrations, "migrate_to_version_%s" % version_slug, None)
+                migrate_fn = getattr(migrations, "migrate_to_version_%s" % version_slug, None)
                 if migrate_fn:
                     flows = []
                     for json_flow in exported_json.get("flows", []):
@@ -3662,7 +3231,7 @@ class FlowRevision(SmartModel):
 
     @classmethod
     def migrate_definition(cls, json_flow, flow, to_version=None):
-        from temba.flows import flow_migrations
+        from temba.flows.legacy import migrations
 
         if not to_version:
             to_version = Flow.FINAL_LEGACY_VERSION
@@ -3670,7 +3239,7 @@ class FlowRevision(SmartModel):
         versions = Flow.get_versions_after(json_flow[Flow.VERSION])
         for version in versions:
             version_slug = version.replace(".", "_")
-            migrate_fn = getattr(flow_migrations, "migrate_to_version_%s" % version_slug, None)
+            migrate_fn = getattr(migrations, "migrate_to_version_%s" % version_slug, None)
 
             if migrate_fn:
                 json_flow = migrate_fn(json_flow, flow)
@@ -5296,8 +4865,14 @@ class TriggerFlowAction(VariableContactAction):
             # start our contacts down the flow
             # our extra will be our flow variables in our message context
             extra = context.get("extra", dict())
-            child_runs = self.flow.start(
-                groups, contacts, restart_participants=True, started_flows=[run.flow.pk], extra=extra, parent_run=run
+            child_runs = legacy.flow_start(
+                self.flow,
+                groups,
+                contacts,
+                restart_participants=True,
+                started_flows=[run.flow.pk],
+                extra=extra,
+                parent_run=run,
             )
 
             # build up all the msgs that where sent by our flow
@@ -5383,8 +4958,14 @@ class StartFlowAction(Action):
         # our extra will be our flow variables in our message context
         extra = context.get("extra", dict())
 
-        child_runs = self.flow.start(
-            [], [run.contact], started_flows=started_flows, restart_participants=True, extra=extra, parent_run=run
+        child_runs = legacy.flow_start(
+            self.flow,
+            [],
+            [run.contact],
+            started_flows=started_flows,
+            restart_participants=True,
+            extra=extra,
+            parent_run=run,
         )
         for run in child_runs:
             for msg in run.start_msgs:
