+++ temba/flows/tests.py
@@ -48,27 +48,6 @@ from temba.values.constants import Value
 
 from . import legacy
 from .checks import mailroom_url
-from .flow_migrations import (
-    map_actions,
-    migrate_export_to_version_9,
-    migrate_to_version_5,
-    migrate_to_version_6,
-    migrate_to_version_7,
-    migrate_to_version_8,
-    migrate_to_version_9,
-    migrate_to_version_10_2,
-    migrate_to_version_10_4,
-    migrate_to_version_11_0,
-    migrate_to_version_11_1,
-    migrate_to_version_11_2,
-    migrate_to_version_11_5,
-    migrate_to_version_11_6,
-    migrate_to_version_11_7,
-    migrate_to_version_11_8,
-    migrate_to_version_11_9,
-    migrate_to_version_11_11,
-    migrate_to_version_11_12,
-)
 from .models import (
     Action,
     ActionSet,
@@ -111,7 +90,6 @@ from .models import (
     HasEmailTest,
     HasStateTest,
     HasWardTest,
-    InGroupTest,
     LteTest,
     LtTest,
     NotEmptyTest,
@@ -136,30 +114,6 @@ from .models import (
 )
 from .tasks import squash_flowpathcounts, squash_flowruncounts, trim_flow_sessions, update_run_expirations_task
 from .views import FlowCRUDL
-
-
-def get_legacy_groups(definition):
-    groups = {}
-    for actionset in definition[Flow.ACTION_SETS]:
-        for action in actionset[Flow.ACTIONS]:
-            for group in action.get("groups", []):
-                groups[group["uuid"]] = group["name"]
-
-    for ruleset in definition[Flow.RULE_SETS]:
-        for rule in ruleset.get(Flow.RULES, []):
-            if rule["test"]["type"] == InGroupTest.TYPE:
-                group = rule["test"]["test"]
-                groups[group["uuid"]] = group["name"]
-    return groups
-
-
-def get_labels(definition):
-    labels = {}
-    for actionset in definition[Flow.ACTION_SETS]:
-        for action in actionset[Flow.ACTIONS]:
-            for label in action.get("labels", []):
-                labels[label["uuid"]] = label["name"]
-    return labels
 
 
 class FlowTest(TembaTest):
@@ -223,7 +177,7 @@ class FlowTest(TembaTest):
 
     @uses_legacy_engine
     def test_archive_interrupt_runs(self):
-        self.flow.start([], [self.contact, self.contact2])
+        legacy.flow_start(self.flow, [], [self.contact, self.contact2])
         self.assertEqual(self.flow.runs.filter(exit_type=None).count(), 2)
 
         self.flow.archive()
@@ -875,7 +829,7 @@ class FlowTest(TembaTest):
         )
 
         # start the flow
-        self.flow.start([], [self.contact, self.contact2])
+        legacy.flow_start(self.flow, [], [self.contact, self.contact2])
 
         # test our stats again
         self.assertEqual(
@@ -951,7 +905,7 @@ class FlowTest(TembaTest):
 
         # create and send a reply
         incoming = self.create_msg(direction=INCOMING, contact=self.contact, text="Orange")
-        self.assertFalse(Flow.find_and_handle(incoming)[0])
+        self.assertFalse(legacy.find_and_handle(incoming)[0])
 
         # no reply, our flow isn't active
         self.assertFalse(Msg.objects.filter(response_to=incoming))
@@ -966,7 +920,7 @@ class FlowTest(TembaTest):
 
         # simulate a response from contact #1
         incoming = self.create_msg(direction=INCOMING, contact=self.contact, text="orange")
-        self.assertTrue(Flow.find_and_handle(incoming)[0])
+        self.assertTrue(legacy.find_and_handle(incoming)[0])
 
         # our message should have gotten a reply
         reply = Msg.objects.get(response_to=incoming)
@@ -995,7 +949,7 @@ class FlowTest(TembaTest):
 
         # at this point there are no more steps to take in the flow, so we shouldn't match anymore
         extra = self.create_msg(direction=INCOMING, contact=self.contact, text="Hello ther")
-        self.assertFalse(Flow.find_and_handle(extra)[0])
+        self.assertFalse(legacy.find_and_handle(extra)[0])
 
         self.assertEqual(
             contact1_run.path,
@@ -1105,7 +1059,7 @@ class FlowTest(TembaTest):
 
     @uses_legacy_engine
     def test_add_messages(self):
-        run, = self.flow.start([], [self.contact])
+        run, = legacy.flow_start(self.flow, [], [self.contact])
 
         msgs = run.get_messages().order_by("id")
         self.assertEqual(len(run.get_msg_events()), 1)
@@ -1131,10 +1085,10 @@ class FlowTest(TembaTest):
         self.org.is_anon = True
         self.org.save()
 
-        (run1,) = self.flow.start([], [self.contact])
+        (run1,) = legacy.flow_start(self.flow, [], [self.contact])
 
         msg = self.create_msg(direction=INCOMING, contact=self.contact, text="orange")
-        Flow.find_and_handle(msg)
+        legacy.find_and_handle(msg)
 
         run1.refresh_from_db()
 
@@ -1179,8 +1133,12 @@ class FlowTest(TembaTest):
         self.login(self.admin)
 
         flow = self.get_flow("two_in_row")
-        contact1_run1, contact2_run1, contact3_run1 = flow.start([], [self.contact, self.contact2, self.contact3])
-        contact1_run2, contact2_run2 = flow.start([], [self.contact, self.contact2], restart_participants=True)
+        contact1_run1, contact2_run1, contact3_run1 = legacy.flow_start(
+            flow, [], [self.contact, self.contact2, self.contact3]
+        )
+        contact1_run2, contact2_run2 = legacy.flow_start(
+            flow, [], [self.contact, self.contact2], restart_participants=True
+        )
 
         for run in (contact1_run1, contact2_run1, contact3_run1, contact1_run2, contact2_run2):
             run.refresh_from_db()
@@ -1454,14 +1412,14 @@ class FlowTest(TembaTest):
 
         favorites = self.get_flow("favorites")
 
-        contact1_run1, contact3_run1 = favorites.start([], [self.contact, self.contact3])
+        contact1_run1, contact3_run1 = legacy.flow_start(favorites, [], [self.contact, self.contact3])
 
         # simulate two runs each for two contacts...
         contact1_in1 = self.create_msg(direction=INCOMING, contact=self.contact, text="light beige")
-        Flow.find_and_handle(contact1_in1)
+        legacy.find_and_handle(contact1_in1)
 
         contact1_in2 = self.create_msg(direction=INCOMING, contact=self.contact, text="red")
-        Flow.find_and_handle(contact1_in2)
+        legacy.find_and_handle(contact1_in2)
 
         # now remap the uuid for our color
         flow_json = favorites.as_json()
@@ -1469,15 +1427,17 @@ class FlowTest(TembaTest):
         flow_json = json.loads(json.dumps(flow_json).replace(color_ruleset["uuid"], str(uuid4())))
         favorites.update(flow_json)
 
-        contact2_run1 = favorites.start([], [self.contact2])[0]
+        contact2_run1 = legacy.flow_start(favorites, [], [self.contact2])[0]
 
         contact2_in1 = self.create_msg(direction=INCOMING, contact=self.contact2, text="green")
-        Flow.find_and_handle(contact2_in1)
+        legacy.find_and_handle(contact2_in1)
 
-        contact1_run2, contact2_run2 = favorites.start([], [self.contact, self.contact2], restart_participants=True)
+        contact1_run2, contact2_run2 = legacy.flow_start(
+            favorites, [], [self.contact, self.contact2], restart_participants=True
+        )
 
         contact1_in3 = self.create_msg(direction=INCOMING, contact=self.contact, text=" blue ")
-        Flow.find_and_handle(contact1_in3)
+        legacy.find_and_handle(contact1_in3)
 
         for run in (contact1_run1, contact2_run1, contact3_run1, contact1_run2, contact2_run2):
             run.refresh_from_db()
@@ -1736,22 +1696,26 @@ class FlowTest(TembaTest):
         self.contact3.name = "Nor\02bert"
         self.contact3.save(update_fields=("name",), handle_update=False)
 
-        contact1_run1, contact2_run1, contact3_run1 = self.flow.start([], [self.contact, self.contact2, self.contact3])
+        contact1_run1, contact2_run1, contact3_run1 = legacy.flow_start(
+            self.flow, [], [self.contact, self.contact2, self.contact3]
+        )
 
         # simulate two runs each for two contacts...
         contact1_in1 = self.create_msg(direction=INCOMING, contact=self.contact, text="light beige")
-        Flow.find_and_handle(contact1_in1)
+        legacy.find_and_handle(contact1_in1)
 
         contact1_in2 = self.create_msg(direction=INCOMING, contact=self.contact, text="orange")
-        Flow.find_and_handle(contact1_in2)
+        legacy.find_and_handle(contact1_in2)
 
         contact2_in1 = self.create_msg(direction=INCOMING, contact=self.contact2, text="green")
-        Flow.find_and_handle(contact2_in1)
+        legacy.find_and_handle(contact2_in1)
 
-        contact1_run2, contact2_run2 = self.flow.start([], [self.contact, self.contact2], restart_participants=True)
+        contact1_run2, contact2_run2 = legacy.flow_start(
+            self.flow, [], [self.contact, self.contact2], restart_participants=True
+        )
 
         contact1_in3 = self.create_msg(direction=INCOMING, contact=self.contact, text=" blue ")
-        Flow.find_and_handle(contact1_in3)
+        legacy.find_and_handle(contact1_in3)
 
         # check can't export anonymously
         exported = self.client.get(reverse("flows.flow_export_results") + "?ids=%d" % self.flow.pk)
@@ -2167,10 +2131,10 @@ class FlowTest(TembaTest):
 
     @uses_legacy_engine
     def test_export_results_remove_control_characters(self):
-        contact1_run1 = self.flow.start([], [self.contact])[0]
+        contact1_run1 = legacy.flow_start(self.flow, [], [self.contact])[0]
 
         contact1_in1 = self.create_msg(direction=INCOMING, contact=self.contact, text="ngert\x07in.")
-        Flow.find_and_handle(contact1_in1)
+        legacy.find_and_handle(contact1_in1)
 
         contact1_run1.refresh_from_db()
 
@@ -2199,9 +2163,9 @@ class FlowTest(TembaTest):
 
     @uses_legacy_engine
     def test_run_as_archive_json(self):
-        contact1_run = self.flow.start([], [self.contact])[0]
+        contact1_run = legacy.flow_start(self.flow, [], [self.contact])[0]
         contact1_in1 = self.create_msg(direction=INCOMING, contact=self.contact, text="green")
-        Flow.find_and_handle(contact1_in1)
+        legacy.find_and_handle(contact1_in1)
 
         # we now have 4 runs in this order of modified_on
         contact1_run.refresh_from_db()
@@ -2302,17 +2266,17 @@ class FlowTest(TembaTest):
 
     @uses_legacy_engine
     def test_export_results_from_archives(self):
-        contact1_run, contact2_run = self.flow.start([], [self.contact, self.contact2])
+        contact1_run, contact2_run = legacy.flow_start(self.flow, [], [self.contact, self.contact2])
         contact1_in1 = self.create_msg(direction=INCOMING, contact=self.contact, text="green")
-        Flow.find_and_handle(contact1_in1)
+        legacy.find_and_handle(contact1_in1)
         contact2_in1 = self.create_msg(direction=INCOMING, contact=self.contact2, text="blue")
-        Flow.find_and_handle(contact2_in1)
+        legacy.find_and_handle(contact2_in1)
 
         # and a run for a different flow
         flow2 = self.get_flow("favorites")
-        contact2_other_flow, = flow2.start([], [self.contact2])
+        contact2_other_flow, = legacy.flow_start(flow2, [], [self.contact2])
 
-        contact3_run, = self.flow.start([], [self.contact3])
+        contact3_run, = legacy.flow_start(self.flow, [], [self.contact3])
 
         # we now have 4 runs in this order of modified_on
         contact1_run.refresh_from_db()
@@ -2423,7 +2387,7 @@ class FlowTest(TembaTest):
     def test_export_results_with_surveyor_msgs(self):
         self.flow.flow_type = Flow.TYPE_SURVEY
         self.flow.save()
-        run = self.flow.start([], [self.contact])[0]
+        run = legacy.flow_start(self.flow, [], [self.contact])[0]
 
         # no urn or channel
         in1 = Msg.create_incoming(None, None, "blue", org=self.org, contact=self.contact)
@@ -2623,7 +2587,7 @@ class FlowTest(TembaTest):
         self.flow.update(json_flow)
 
         self.mockRequest("POST", "/coupon", '{"coupon": "NEXUS4"}')
-        run, = self.flow.start([], [self.contact])
+        run, = legacy.flow_start(self.flow, [], [self.contact])
 
         self.assertTrue(run.path)
         self.assertTrue(Msg.objects.all())
@@ -3416,9 +3380,9 @@ class FlowTest(TembaTest):
         color_ruleset.save()
 
         # start the flow
-        self.flow.start([], [self.contact])
+        legacy.flow_start(self.flow, [], [self.contact])
         sms = self.create_msg(direction=INCOMING, contact=self.contact, text="My answer is 15")
-        self.assertTrue(Flow.find_and_handle(sms)[0])
+        self.assertTrue(legacy.find_and_handle(sms)[0])
 
         # get our run and assert our value is saved (as a string)
         run = FlowRun.objects.get(flow=self.flow, contact=self.contact)
@@ -3438,9 +3402,9 @@ class FlowTest(TembaTest):
 
     @uses_legacy_engine
     def test_null_categories(self):
-        self.flow.start([], [self.contact])
+        legacy.flow_start(self.flow, [], [self.contact])
         sms = self.create_msg(direction=INCOMING, contact=self.contact, text="blue")
-        self.assertTrue(Flow.find_and_handle(sms)[0])
+        self.assertTrue(legacy.find_and_handle(sms)[0])
 
         FlowCategoryCount.objects.get(category_name="Blue", result_name="color", result_key="color", count=1)
 
@@ -3975,7 +3939,7 @@ class FlowTest(TembaTest):
         )  # by saved_on
 
         # start a contact in a flow
-        self.flow.start([], [self.contact])
+        legacy.flow_start(self.flow, [], [self.contact])
 
         # test getting the json
         response = self.client.get(reverse("flows.flow_json", args=[self.flow.uuid]))
@@ -4289,7 +4253,7 @@ class FlowTest(TembaTest):
     @uses_legacy_engine
     def test_flow_start_with_start_msg(self):
         msg_in = self.create_msg(direction=INCOMING, contact=self.contact, text="I am coming")
-        run, = self.flow.start([], [self.contact], start_msg=msg_in)
+        run, = legacy.flow_start(self.flow, [], [self.contact], start_msg=msg_in)
 
         msg_in.refresh_from_db()
         msg_out = Msg.objects.get(direction="O")
@@ -4306,7 +4270,7 @@ class FlowTest(TembaTest):
     @uses_legacy_engine
     def test_quick_replies(self):
         flow = self.get_flow("quick_replies")
-        run, = flow.start([], [self.contact4])
+        run, = legacy.flow_start(flow, [], [self.contact4])
 
         # contact language is Portugese but this isn't an org language so we should use English
         msg = Msg.objects.filter(direction="O").last()
@@ -4314,18 +4278,18 @@ class FlowTest(TembaTest):
 
         # add Portugese as an org language and try again
         self.org.set_languages(self.admin, ["eng", "por"], "eng")
-        run, = flow.start([], [self.contact4], restart_participants=True)
+        run, = legacy.flow_start(flow, [], [self.contact4], restart_participants=True)
 
         msg = Msg.objects.filter(direction="O").last()
         self.assertEqual(msg.metadata, {"quick_replies": ["Sim", "No"]})
 
     @uses_legacy_engine
     def test_multiple(self):
-        run1, = self.flow.start([], [self.contact])
+        run1, = legacy.flow_start(self.flow, [], [self.contact])
 
         # create a second flow and start our same contact
         self.flow2 = self.flow.copy(self.flow, self.flow.created_by)
-        run2, = self.flow2.start([], [self.contact])
+        run2, = legacy.flow_start(self.flow2, [], [self.contact])
 
         run1.refresh_from_db()
         run2.refresh_from_db()
@@ -4339,7 +4303,7 @@ class FlowTest(TembaTest):
 
         # send in a message
         incoming = self.create_msg(direction=INCOMING, contact=self.contact, text="Orange", created_on=timezone.now())
-        self.assertTrue(Flow.find_and_handle(incoming)[0])
+        self.assertTrue(legacy.find_and_handle(incoming)[0])
 
         run1.refresh_from_db()
         run2.refresh_from_db()
@@ -4349,7 +4313,7 @@ class FlowTest(TembaTest):
         self.assertEqual(len(run2.path), 3)
 
         # start the flow again for our contact
-        run3, = self.flow.start([], [self.contact], restart_participants=True)
+        run3, = legacy.flow_start(self.flow, [], [self.contact], restart_participants=True)
 
         run1.refresh_from_db()
         run3.refresh_from_db()
@@ -4363,7 +4327,7 @@ class FlowTest(TembaTest):
 
         # send in a message, this should be handled by our first flow, which has a more recent run active
         incoming = self.create_msg(direction=INCOMING, contact=self.contact, text="blue")
-        self.assertTrue(Flow.find_and_handle(incoming)[0])
+        self.assertTrue(legacy.find_and_handle(incoming)[0])
 
         run1.refresh_from_db()
         run3.refresh_from_db()
@@ -4372,7 +4336,7 @@ class FlowTest(TembaTest):
         self.assertEqual(len(run3.path), 3)
 
         # if we exclude existing and try starting again, nothing happens
-        self.flow.start([], [self.contact], restart_participants=False)
+        legacy.flow_start(self.flow, [], [self.contact], restart_participants=False)
 
         # no new runs
         self.assertEqual(self.flow.runs.count(), 2)
@@ -4900,7 +4864,7 @@ class ActionTest(TembaTest):
         self.flow.name = "Parent"
         self.flow.save()
 
-        self.flow.start([], [self.contact])
+        legacy.flow_start(self.flow, [], [self.contact])
 
         sms = Msg.create_incoming(self.channel, "tel:+250788382382", "Blue is my favorite")
 
@@ -5230,12 +5194,12 @@ class FlowRunTest(TembaTest):
 
     @uses_legacy_engine
     def test_is_completed(self):
-        self.flow.start([], [self.contact])
+        legacy.flow_start(self.flow, [], [self.contact])
 
         self.assertFalse(FlowRun.objects.get(contact=self.contact).is_completed())
 
         incoming = self.create_msg(direction=INCOMING, contact=self.contact, text="orange")
-        Flow.find_and_handle(incoming)
+        legacy.find_and_handle(incoming)
 
         self.assertTrue(FlowRun.objects.get(contact=self.contact).is_completed())
 
@@ -5424,7 +5388,7 @@ class WebhookTest(TembaTest):
         flow = self.get_flow("triggered", substitutions)
 
         self.mockRequest("GET", "/where", '{ "text": "(I came from a webhook)" }')
-        flow.start(groups=[], contacts=[contact1], restart_participants=True)
+        legacy.flow_start(flow, groups=[], contacts=[contact1], restart_participants=True)
 
         # first message from our trigger flow action
         msg = Msg.objects.all().order_by("-created_on")[0]
@@ -5445,7 +5409,7 @@ class WebhookTest(TembaTest):
 
         self.mockRequest("GET", "/", '{ "decimal": 0.1 }')
 
-        run1, = flow.start([], [contact])
+        run1, = legacy.flow_start(flow, [], [contact])
         run1.refresh_from_db()
 
         self.assertEqual(run1.fields, {"decimal": Decimal("0.1"), "webhook": '{ "decimal": 0.1 }'})
@@ -5459,7 +5423,7 @@ class WebhookTest(TembaTest):
 
         self.mockRequest("GET", "/check_order.php?phone=%2B250788383383", '{ "text": "Get", "blank": "" }')
 
-        run1, = flow.start([], [contact])
+        run1, = legacy.flow_start(flow, [], [contact])
         run1.refresh_from_db()
 
         self.assertEqual(run1.fields, {"text": "Get", "blank": "", "response_1": '{ "text": "Get", "blank": "" }'})
@@ -5494,7 +5458,7 @@ class WebhookTest(TembaTest):
 
         self.mockRequest("POST", "/check_order.php?phone=%2B250788383383", '{ "text": "Post", "blank": "" }')
 
-        run2, = flow.start([], [contact], restart_participants=True)
+        run2, = legacy.flow_start(flow, [], [contact], restart_participants=True)
         run2.refresh_from_db()
 
         self.assertEqual(
@@ -5522,7 +5486,7 @@ class WebhookTest(TembaTest):
         # check parsing of a JSON array response from a webhook
         self.mockRequest("POST", "/check_order.php?phone=%2B250788383383", '["zero", "one", "two"]')
 
-        run3, = flow.start([], [contact], restart_participants=True)
+        run3, = legacy.flow_start(flow, [], [contact], restart_participants=True)
         run3.refresh_from_db()
 
         self.assertEqual(run3.fields, {"0": "zero", "1": "one", "2": "two", "response_1": '["zero", "one", "two"]'})
@@ -5537,7 +5501,7 @@ class WebhookTest(TembaTest):
         body = json.dumps(["x"] * 300)
         self.mockRequest("POST", "/check_order.php?phone=%2B250788383383", body)
 
-        run4, = flow.start([], [contact], restart_participants=True)
+        run4, = legacy.flow_start(flow, [], [contact], restart_participants=True)
         run4.refresh_from_db()
         expected_fields = {str(n): "x" for n in range(256)}
         expected_fields["response_1"] = body[: Value.MAX_VALUE_LEN]
@@ -5546,14 +5510,14 @@ class WebhookTest(TembaTest):
         # check we handle a non-dict or list response
         self.mockRequest("POST", "/check_order.php?phone=%2B250788383383", "12345")
 
-        run5, = flow.start([], [contact], restart_participants=True)
+        run5, = legacy.flow_start(flow, [], [contact], restart_participants=True)
         run5.refresh_from_db()
         self.assertEqual(run5.fields, {"response_1": "12345"})
 
         # check we handle a non-JSON response
         self.mockRequest("POST", "/check_order.php?phone=%2B250788383383", "asdfasdfasdf")
 
-        run6, = flow.start([], [contact], restart_participants=True)
+        run6, = legacy.flow_start(flow, [], [contact], restart_participants=True)
         run6.refresh_from_db()
 
         self.assertEqual(run6.fields, {"response_1": "asdfasdfasdf"})
@@ -5567,7 +5531,7 @@ class WebhookTest(TembaTest):
         # check a webhook that responds with a 500 error
         self.mockRequest("POST", "/check_order.php?phone=%2B250788383383", "Server Error", status=500)
 
-        run7, = flow.start([], [contact], restart_participants=True)
+        run7, = legacy.flow_start(flow, [], [contact], restart_participants=True)
         run7.refresh_from_db()
 
         self.assertEqual(run7.fields, {"response_1": "Server Error"})
@@ -5589,7 +5553,7 @@ class WebhookTest(TembaTest):
         body = '{ "text": "Valid", "error": "400", "message": "Missing field in request" }'
         self.mockRequest("POST", "/check_order.php?phone=%2B250788383383", body, status=400)
 
-        run8, = flow.start([], [contact], restart_participants=True)
+        run8, = legacy.flow_start(flow, [], [contact], restart_participants=True)
         run8.refresh_from_db()
 
         self.assertEqual(
@@ -5612,7 +5576,7 @@ class WebhookTest(TembaTest):
         webhook_flow = self.get_flow("resthooks")
 
         # we don't have the resthook registered yet, so this won't trigger any calls
-        webhook_flow.start([], [self.contact])
+        legacy.flow_start(webhook_flow, [], [self.contact])
 
         # should have two messages
         msgs = list(self.contact.msgs.order_by("id"))
@@ -5640,7 +5604,7 @@ class WebhookTest(TembaTest):
         self.mockRequest("POST", "/bar", "Failure 2", status=400)
 
         # start over
-        webhook_flow.start([], [self.contact], restart_participants=True)
+        legacy.flow_start(webhook_flow, [], [self.contact], restart_participants=True)
 
         msgs = list(self.contact.msgs.order_by("id"))
 
@@ -5656,7 +5620,7 @@ class WebhookTest(TembaTest):
         self.mockRequest("POST", "/foo", '{ "code": "XXCXCVXVVXV" }')
 
         # start over
-        webhook_flow.start([], [self.contact], restart_participants=True)
+        legacy.flow_start(webhook_flow, [], [self.contact], restart_participants=True)
 
         # we should also have unsubscribed from one of our endpoints
         self.assertTrue(resthook.subscribers.filter(is_active=True, target_url="http://localhost:49999/foo"))
@@ -5872,7 +5836,7 @@ class FlowsTest(FlowFileTest):
         rule_set1, rule_set2 = favorites.rule_sets.order_by("y")[:2]
         red_rule = rule_set1.rules[0]
 
-        run, = favorites.start([], [self.contact])
+        run, = legacy.flow_start(favorites, [], [self.contact])
 
         msg1 = Msg.objects.get()
         self.assertEqual(msg1.direction, "O")
@@ -6037,7 +6001,7 @@ class FlowsTest(FlowFileTest):
         rule_set1, rule_set2 = favorites.rule_sets.order_by("y")[:2]
         navy_rule = rule_set1.rules[3]
 
-        run, = favorites.start([], [self.contact])
+        run, = legacy.flow_start(favorites, [], [self.contact])
         Msg.create_incoming(self.channel, "tel:+12065552020", "navy")
 
         run.refresh_from_db()
@@ -6079,7 +6043,7 @@ class FlowsTest(FlowFileTest):
         rule_set1, rule_set2 = flow.rule_sets.order_by("y")
         ben_rule = rule_set2.rules[0]
 
-        run, = flow.start([], [self.contact])
+        run, = legacy.flow_start(flow, [], [self.contact])
 
         # answer with A to first ruleset which is Action (A) or Ruleset (R)
         Msg.create_incoming(self.channel, "tel:+12065552020", "A")
@@ -6089,7 +6053,7 @@ class FlowsTest(FlowFileTest):
         self.assertEqual(run.exit_type, "C")
         self.assertEqual(run.path[2]["exit_uuid"], str(action_set2.exit_uuid))
 
-        run, = flow.start([], [self.contact], restart_participants=True)
+        run, = legacy.flow_start(flow, [], [self.contact], restart_participants=True)
 
         # this time choose to end on a non-waiting ruleset
         Msg.create_incoming(self.channel, "tel:+12065552020", "R")
@@ -6102,7 +6066,7 @@ class FlowsTest(FlowFileTest):
     @uses_legacy_engine
     def test_resuming_run_with_old_uuidless_message(self):
         favorites = self.get_flow("favorites")
-        run, = favorites.start([], [self.contact])
+        run, = legacy.flow_start(favorites, [], [self.contact])
 
         Msg.create_incoming(self.channel, "tel:+12065552020", "I like red")
 
@@ -6139,7 +6103,7 @@ class FlowsTest(FlowFileTest):
         # if there is no urn, it still works, but is omitted
         empty_post = self.mockRequest("POST", "/send_results", '{"received":"ruleset"}', content_type=ctype)
         empty = self.create_contact("Empty Contact")
-        flow.start([], [empty])
+        legacy.flow_start(flow, [], [empty])
         self.send("39", empty)
         self.send("tornado", empty)
         self.assertIsNone(empty_post.data["contact"].get("urn"))
@@ -6147,7 +6111,7 @@ class FlowsTest(FlowFileTest):
         # test fallback urn
         fallback_post = self.mockRequest("POST", "/send_results", '{"received":"ruleset"}', content_type=ctype)
         empty_flow = self.get_flow("empty_payload")
-        empty_flow.start([], [self.contact], restart_participants=True)
+        legacy.flow_start(empty_flow, [], [self.contact], restart_participants=True)
         self.assertEqual("tel:+12065552020", fallback_post.data["contact"]["urn"])
 
         def assert_payload(payload, path_length, result_count, results):
@@ -6210,7 +6174,7 @@ class FlowsTest(FlowFileTest):
     @uses_legacy_engine
     def test_send_msg_to_urnless(self):
         flow = self.get_flow("send_msg_to_urnless")
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         # check we create the contact
         self.assertIsNotNone(Contact.objects.get(name="Bob"))
@@ -6769,7 +6733,7 @@ class FlowsTest(FlowFileTest):
         flow = self.get_flow("send_all")
 
         contact = self.create_contact("Stephen", "+12078778899", twitter="stephen")
-        flow.start(groups=[], contacts=[contact], restart_participants=True)
+        legacy.flow_start(flow, groups=[], contacts=[contact], restart_participants=True)
 
         replies = Msg.objects.filter(contact=contact, direction="O")
         self.assertEqual(replies.count(), 1)
@@ -6783,7 +6747,7 @@ class FlowsTest(FlowFileTest):
         Channel.create(self.org, self.user, None, "TT")
         flow.org.clear_cached_schemes()
 
-        flow.start(groups=[], contacts=[contact], restart_participants=True)
+        legacy.flow_start(flow, groups=[], contacts=[contact], restart_participants=True)
 
         replies = Msg.objects.filter(contact=contact, direction="O")
         self.assertEqual(replies.count(), 2)
@@ -6794,7 +6758,7 @@ class FlowsTest(FlowFileTest):
         Msg.objects.all().delete()
 
         flow = self.get_flow("two_to_all")
-        flow.start(groups=[], contacts=[contact], restart_participants=True)
+        legacy.flow_start(flow, groups=[], contacts=[contact], restart_participants=True)
 
         replies = Msg.objects.filter(contact=contact, direction="O")
         self.assertEqual(replies.count(), 4)
@@ -6826,7 +6790,7 @@ class FlowsTest(FlowFileTest):
         # no params returns no results
         assert_recent(self.client.get(recent_messages_url), [])
 
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
         self.create_msg(direction=INCOMING, contact=self.contact, text="chartreuse").handle()
 
         response = self.client.get(recent_messages_url + entry_params)
@@ -7005,7 +6969,7 @@ class FlowsTest(FlowFileTest):
         color_blue_uuid = rules[-4].uuid
 
         # we don't know this shade of green, it should route us to the beginning again
-        run1, = flow.start([], [self.contact])
+        run1, = legacy.flow_start(flow, [], [self.contact])
         self.create_msg(direction=INCOMING, contact=self.contact, text="chartreuse").handle()
 
         (active, visited) = flow.get_activity()
@@ -7337,7 +7301,7 @@ class FlowsTest(FlowFileTest):
         blue_rule = rules[1]
         other_rule = rules[2]
 
-        run, = flow.start([], [self.contact])
+        run, = legacy.flow_start(flow, [], [self.contact])
 
         now = datetime.datetime(2014, 1, 2, 3, 4, 5, 6, timezone.utc)
 
@@ -7443,7 +7407,7 @@ class FlowsTest(FlowFileTest):
 
         # send a message, no flow should handle us since we are done
         incoming = self.create_msg(direction=INCOMING, contact=self.contact, text="Unhandled")
-        handled = Flow.find_and_handle(incoming)[0]
+        handled = legacy.find_and_handle(incoming)[0]
         self.assertFalse(handled)
 
         # now wire up our finished action to the start of our flow
@@ -7586,7 +7550,7 @@ class FlowsTest(FlowFileTest):
         flow = self.get_flow("non_blocking_rule_first")
 
         eminem = self.create_contact("Eminem", "+12065550100")
-        flow.start(groups=[], contacts=[eminem])
+        legacy.flow_start(flow, groups=[], contacts=[eminem])
         msg = Msg.objects.filter(direction="O", contact=eminem).first()
         self.assertEqual("Hi there Eminem", msg.text)
 
@@ -7596,7 +7560,7 @@ class FlowsTest(FlowFileTest):
         ruleset.save()
 
         tupac = self.create_contact("Tupac", "+12065550101")
-        flow.start(groups=[], contacts=[tupac])
+        legacy.flow_start(flow, groups=[], contacts=[tupac])
         msg = Msg.objects.filter(direction="O", contact=tupac).first()
         self.assertEqual("Hi there Tupac", msg.text)
 
@@ -7604,7 +7568,7 @@ class FlowsTest(FlowFileTest):
     def test_webhook_rule_first(self):
         flow = self.get_flow("webhook_rule_first")
         tupac = self.create_contact("Tupac", "+12065550101")
-        flow.start(groups=[], contacts=[tupac])
+        legacy.flow_start(flow, groups=[], contacts=[tupac])
 
         # a message should have been sent
         msg = Msg.objects.filter(direction="O", contact=tupac).first()
@@ -7781,7 +7745,7 @@ class FlowsTest(FlowFileTest):
                     group_count += 1
         self.assertEqual(2, group_count)
 
-        run, = flow.start([], [self.contact])
+        run, = legacy.flow_start(flow, [], [self.contact])
 
         # not in any group
         self.assertEqual(0, ContactGroup.user_groups.filter(contacts__in=[self.contact]).count())
@@ -7839,7 +7803,7 @@ class FlowsTest(FlowFileTest):
 
         # if contact has null name, value will be empty string
         nameless = self.create_contact(name=None, number="+12065553030")
-        run, = flow.start([], [nameless])
+        run, = legacy.flow_start(flow, [], [nameless])
 
         self.send("split", contact=nameless)
         self.assertEqual(
@@ -7864,7 +7828,7 @@ class FlowsTest(FlowFileTest):
     def test_media_first_action(self):
         flow = self.get_flow("media_first_action")
 
-        runs = flow.start([], [self.contact])
+        runs = legacy.flow_start(flow, [], [self.contact])
         self.assertEqual(1, len(runs))
 
         msg = self.contact.msgs.get()
@@ -7880,7 +7844,7 @@ class FlowsTest(FlowFileTest):
         self.contact.name = "Ben Haggerty"
         self.contact.save(update_fields=("name",), handle_update=False)
 
-        runs = flow.start([], [self.contact])
+        runs = legacy.flow_start(flow, [], [self.contact])
         self.assertEqual(1, len(runs))
         self.assertEqual(self.contact.msgs.get().text, "Hi Ben Haggerty, what is your phone number?")
 
@@ -7925,7 +7889,7 @@ class FlowsTest(FlowFileTest):
         rule_flow = self.get_flow("group_rule_first")
 
         # start our contact down it
-        rule_flow.start([], [self.contact], restart_participants=True)
+        legacy.flow_start(rule_flow, [], [self.contact], restart_participants=True)
 
         # contact should get a message that they didn't match either group
         self.assertLastResponse("You are something else.")
@@ -7933,7 +7897,7 @@ class FlowsTest(FlowFileTest):
         # add them to the father's group
         self.create_group("Fathers", [self.contact])
 
-        rule_flow.start([], [self.contact], restart_participants=True)
+        legacy.flow_start(rule_flow, [], [self.contact], restart_participants=True)
         self.assertLastResponse("You are a father.")
 
     @uses_legacy_engine
@@ -8278,7 +8242,7 @@ class FlowsTest(FlowFileTest):
     @uses_legacy_engine
     def test_initial_expiration(self):
         flow = self.get_flow("favorites")
-        flow.start(groups=[], contacts=[self.contact])
+        legacy.flow_start(flow, groups=[], contacts=[self.contact])
 
         run = FlowRun.objects.get()
         self.assertTrue(run.expires_on)
@@ -8328,7 +8292,7 @@ class FlowsTest(FlowFileTest):
         flow2 = Flow.objects.get(name="Second Flow")
 
         # start the flow, shouldn't get into a loop, but both should get started
-        flow1.start([], [self.contact])
+        legacy.flow_start(flow1, [], [self.contact])
 
         self.assertTrue(FlowRun.objects.get(flow=flow1, contact=self.contact))
         self.assertTrue(FlowRun.objects.get(flow=flow2, contact=self.contact))
@@ -8341,7 +8305,7 @@ class FlowsTest(FlowFileTest):
         flow2 = Flow.objects.all()[0]
 
         # start the flow, should not get into a loop
-        flow1.start([], [self.contact])
+        legacy.flow_start(flow1, [], [self.contact])
 
         self.assertTrue(FlowRun.objects.get(flow=flow1, contact=self.contact))
         self.assertTrue(FlowRun.objects.get(flow=flow2, contact=self.contact))
@@ -8377,7 +8341,7 @@ class FlowsTest(FlowFileTest):
         joe = self.create_contact("joe", "112233")
 
         parent = Flow.objects.get(name="Priority Parent")
-        parent.start([], [self.contact, joe])
+        legacy.flow_start(parent, [], [self.contact, joe])
 
         self.assertEqual(8, Msg.objects.filter(direction="O").count())
 
@@ -8404,7 +8368,7 @@ class FlowsTest(FlowFileTest):
         # contact, but the messages should still be low prioirty
         self.get_flow("priorities")
         parent = Flow.objects.get(name="Priority Parent")
-        parent.start([], [self.contact], restart_participants=True)
+        legacy.flow_start(parent, [], [self.contact], restart_participants=True)
 
         self.assertEqual(4, Msg.objects.count())
         self.assertEqual(0, Broadcast.objects.count())
@@ -8422,8 +8386,11 @@ class FlowsTest(FlowFileTest):
         subflow_ruleset = RuleSet.objects.get(flow=parent, ruleset_type="subflow")
         subflow_reply = ActionSet.objects.get(flow=parent, y=386, x=341)
 
-        parent.start(
-            groups=[], contacts=[self.contact, self.create_contact("joe", "+12347778888")], restart_participants=True
+        legacy.flow_start(
+            parent,
+            groups=[],
+            contacts=[self.contact, self.create_contact("joe", "+12347778888")],
+            restart_participants=True,
         )
 
         msg = Msg.objects.filter(contact=self.contact).first()
@@ -8483,7 +8450,7 @@ class FlowsTest(FlowFileTest):
         self.get_flow("subflow")
         parent = Flow.objects.get(org=self.org, name="Parent Flow")
 
-        parent.start(groups=[], contacts=[self.contact], restart_participants=True)
+        legacy.flow_start(parent, groups=[], contacts=[self.contact], restart_participants=True)
         self.send_message(parent, "color", assert_reply=False)
 
         # we should now have two active flows
@@ -8506,7 +8473,7 @@ class FlowsTest(FlowFileTest):
         self.get_flow("subflow")
         parent = Flow.objects.get(org=self.org, name="Parent Flow")
 
-        parent.start(groups=[], contacts=[self.contact], restart_participants=True)
+        legacy.flow_start(parent, groups=[], contacts=[self.contact], restart_participants=True)
         self.send_message(parent, "color", assert_reply=False)
 
         # we should now have two active flows
@@ -8535,7 +8502,7 @@ class FlowsTest(FlowFileTest):
         self.get_flow("subflow")
         parent = Flow.objects.get(org=self.org, name="Parent Flow")
 
-        parent.start(groups=[], contacts=[self.contact], restart_participants=True)
+        legacy.flow_start(parent, groups=[], contacts=[self.contact], restart_participants=True)
         self.send_message(parent, "color", assert_reply=False)
 
         # we should now have two active flows
@@ -8559,7 +8526,7 @@ class FlowsTest(FlowFileTest):
     def test_subflow_no_interaction(self):
         self.get_flow("subflow_no_pause")
         parent = Flow.objects.get(org=self.org, name="Flow A")
-        parent.start(groups=[], contacts=[self.contact], restart_participants=True)
+        legacy.flow_start(parent, groups=[], contacts=[self.contact], restart_participants=True)
 
         # check we got our three messages, the third populated by the child, but sent form the parent
         msgs = Msg.objects.order_by("created_on")
@@ -8575,7 +8542,7 @@ class FlowsTest(FlowFileTest):
         self.get_flow("subflow_with_startflow")
 
         parent = Flow.objects.get(name="Subflow 1")
-        parent.start(groups=[], contacts=[self.contact])
+        legacy.flow_start(parent, groups=[], contacts=[self.contact])
 
     @uses_legacy_engine
     def test_trigger_flow_complete(self):
@@ -8585,7 +8552,7 @@ class FlowsTest(FlowFileTest):
 
         parent = Flow.objects.get(org=self.org, name="Flow A")
 
-        parent.start(groups=[], contacts=[self.contact], restart_participants=True)
+        legacy.flow_start(parent, groups=[], contacts=[self.contact], restart_participants=True)
 
         self.assertEqual(1, FlowRun.objects.filter(contact=self.contact).count())
         self.assertEqual(1, FlowRun.objects.filter(contact=contact2).count())
@@ -8613,14 +8580,14 @@ class FlowsTest(FlowFileTest):
 
         # start our flow without a message (simulating it being fired by a trigger or the simulator)
         # this will evaluate requires_step() to make sure it handles localized flows
-        runs = flow.start([], [self.contact])
+        runs = legacy.flow_start(flow, [], [self.contact])
         self.assertEqual(1, len(runs))
         self.assertEqual(self.contact.msgs.get().text, "You are not in the enrolled group.")
 
         enrolled_group = ContactGroup.create_static(self.org, self.user, "Enrolled")
         enrolled_group.update_contacts(self.user, [self.contact], True)
 
-        runs_started = flow.start([], [self.contact], restart_participants=True)
+        runs_started = legacy.flow_start(flow, [], [self.contact], restart_participants=True)
         self.assertEqual(1, len(runs_started))
 
         msgs = list(self.contact.msgs.order_by("id"))
@@ -8737,7 +8704,7 @@ class FlowsTest(FlowFileTest):
         with patch("temba.flows.models.AirtimeTransfer.trigger_airtime_event") as mock_trigger_event:
             mock_trigger_event.return_value = airtime_event
 
-            runs = flow.start([], [self.contact])
+            runs = legacy.flow_start(flow, [], [self.contact])
             self.assertEqual(1, len(runs))
             self.assertEqual(self.contact.msgs.get().text, "Message complete")
 
@@ -8746,7 +8713,7 @@ class FlowsTest(FlowFileTest):
 
             mock_trigger_event.return_value = airtime_event
 
-            runs = flow.start([], [self.contact], restart_participants=True)
+            runs = legacy.flow_start(flow, [], [self.contact], restart_participants=True)
             self.assertEqual(1, len(runs))
 
             msgs = list(self.contact.msgs.order_by("id"))
@@ -8772,7 +8739,7 @@ class FlowsTest(FlowFileTest):
         self.org.refresh_transferto_account_currency()
 
         flow = self.get_flow("airtime")
-        runs = flow.start([], [self.contact])
+        runs = legacy.flow_start(flow, [], [self.contact])
         self.assertEqual(1, len(runs))
         self.assertEqual(self.contact.msgs.get().text, "Message complete")
 
@@ -8795,7 +8762,7 @@ class FlowsTest(FlowFileTest):
             MockResponse(200, "error_code=0\r\nerror_txt=\r\n"),
         ]
 
-        runs = flow.start([], [self.contact], restart_participants=True)
+        runs = legacy.flow_start(flow, [], [self.contact], restart_participants=True)
         self.assertEqual(1, len(runs))
         msgs = list(self.contact.msgs.order_by("id"))
         self.assertEqual(msgs[1].text, "Message failed")
@@ -8826,7 +8793,7 @@ class FlowsTest(FlowFileTest):
         contact2 = self.create_contact(name="Bismack Biyombo", number="+250788123123", twitter="biyombo")
         self.assertEqual(contact2.get_urn().path, "biyombo")
 
-        runs = flow.start([], [contact2], restart_participants=True)
+        runs = legacy.flow_start(flow, [], [contact2], restart_participants=True)
         self.assertEqual(1, len(runs))
         self.assertEqual(1, contact2.msgs.all().count())
         self.assertEqual("Message complete", contact2.msgs.all()[0].text)
@@ -8852,7 +8819,7 @@ class FlowsTest(FlowFileTest):
             MockResponse(200, "error_code=0\r\nerror_txt=\r\n"),
         ]
 
-        runs = flow.start([], [self.contact], restart_participants=True)
+        runs = legacy.flow_start(flow, [], [self.contact], restart_participants=True)
         self.assertEqual(1, len(runs))
 
         msgs = list(self.contact.msgs.order_by("id"))
@@ -8925,1314 +8892,6 @@ class FlowsTest(FlowFileTest):
         self.assertEqual(str(run.current_node_uuid), beerRuleSet.uuid)
 
 
-class FlowMigrationTest(FlowFileTest):
-    def test_is_before_version(self):
-
-        # works with numbers
-        self.assertTrue(Flow.is_before_version(5, 6))
-
-        self.assertTrue(Flow.is_before_version("10", "10.1"))
-        self.assertFalse(Flow.is_before_version("10", "9"))
-
-        # unknown versions return false
-        self.assertFalse(Flow.is_before_version("3.1", "5"))
-        self.assertFalse(Flow.is_before_version("200", "5"))
-        self.assertFalse(Flow.is_before_version("3.1", "3.5"))
-
-        self.assertFalse(Flow.is_before_version(Flow.FINAL_LEGACY_VERSION, 10))
-
-    def migrate_flow(self, flow, to_version=None):
-
-        if not to_version:
-            to_version = Flow.FINAL_LEGACY_VERSION
-
-        flow_json = flow.as_json()
-        if Flow.is_before_version(flow.version_number, "6"):
-            revision = flow.revisions.all().order_by("-revision").first()
-            flow_json = dict(
-                definition=flow_json,
-                flow_type=flow.flow_type,
-                expires=flow.expires_after_minutes,
-                id=flow.pk,
-                revision=revision.revision if revision else 1,
-            )
-
-        flow_json = FlowRevision.migrate_definition(flow_json, flow, to_version=to_version)
-        if "definition" in flow_json:
-            flow_json = flow_json["definition"]
-
-        flow.update(flow_json)
-        return Flow.objects.get(pk=flow.pk)
-
-    def test_migrate_with_flow_user(self):
-        flow = Flow.objects.create(
-            name="Favorites",
-            org=self.org,
-            created_by=self.admin,
-            modified_by=self.admin,
-            saved_by=self.admin,
-            version_number="7",
-            flow_type="M",
-        )
-
-        flow_json = self.get_flow_json("favorites")
-        FlowRevision.objects.create(
-            flow=flow, definition=flow_json, spec_version=7, revision=1, created_by=self.admin, modified_by=self.admin
-        )
-
-        old_json = flow.as_json()
-
-        saved_on = flow.saved_on
-        modified_on = flow.modified_on
-        flow.ensure_current_version()
-        flow.refresh_from_db()
-
-        # system migration should not affect our saved_on even tho we are modified
-        self.assertNotEqual(modified_on, flow.modified_on)
-        self.assertEqual(saved_on, flow.saved_on)
-
-        # but should still create a revision using the flow user
-        self.assertEqual(1, flow.revisions.filter(created_by=get_flow_user(self.org)).count())
-
-        # should see the system user on our revision json
-        self.login(self.admin)
-        response = self.client.get(reverse("flows.flow_revisions", args=[flow.uuid]))
-
-        self.assertContains(response, "System Update")
-        self.assertEqual(2, len(response.json()["results"]))
-
-        # attempt to save with old json, no bueno
-        response = self.client.post(
-            reverse("flows.flow_json", args=[flow.uuid]), data=json.dumps(old_json), content_type="application/json"
-        )
-
-        self.assertEqual(response.status_code, 400)
-        self.assertEqual(
-            response.json(),
-            {
-                "description": "rapidpro_flow is currently editing this Flow. Your changes will not be saved until you refresh your browser.",
-                "status": "failure",
-            },
-        )
-
-        # now refresh and save a new version
-        flow.update(flow.as_json(), user=self.admin)
-
-        self.assertEqual(flow.revisions.count(), 3)
-        self.assertEqual(flow.revisions.filter(created_by=get_flow_user(self.org)).count(), 1)
-
-        # if we request a specific revision by id, flow will be migrated to new format
-        revision = flow.revisions.order_by("id").last()
-        response = self.client.get("%s%s/" % (reverse("flows.flow_revisions", args=[flow.uuid]), str(revision.id)))
-
-        self.assertEqual(response.json()["spec_version"], "13.0.0")
-
-    def test_migrate_malformed_single_message_flow(self):
-
-        flow = Flow.objects.create(
-            name="Single Message Flow",
-            org=self.org,
-            created_by=self.admin,
-            modified_by=self.admin,
-            saved_by=self.admin,
-            version_number=3,
-        )
-
-        flow_json = self.get_flow_json("malformed_single_message")["definition"]
-
-        FlowRevision.objects.create(
-            flow=flow, definition=flow_json, spec_version=3, revision=1, created_by=self.admin, modified_by=self.admin
-        )
-
-        flow.ensure_current_version()
-        flow_json = flow.as_json()
-
-        self.assertEqual(len(flow_json["action_sets"]), 1)
-        self.assertEqual(len(flow_json["rule_sets"]), 0)
-        self.assertEqual(flow_json["version"], Flow.FINAL_LEGACY_VERSION)
-        self.assertEqual(flow_json["metadata"]["revision"], 2)
-
-    def test_update_with_ruleset_to_actionset_change(self):
-        flow = self.get_flow("favorites")
-
-        # so that we check clearing of prefetched nodes that might be deleted in .update()
-        flow = Flow.objects.prefetch_related("action_sets", "rule_sets").get(id=flow.id)
-
-        flow_json = flow.as_json()
-
-        # remove first ruleset
-        ruleset1 = flow_json["rule_sets"][0]
-        flow_json["rule_sets"] = flow_json["rule_sets"][1:]
-
-        # create actionset in its place with same uuid etc
-        flow_json["action_sets"].append(
-            {
-                "uuid": ruleset1["uuid"],
-                "actions": [{"uuid": "da6d7657-8940-4778-ada8-27a2035a8352", "type": "lang", "language": "fra"}],
-                "x": ruleset1["x"],
-                "y": ruleset1["y"],
-                "destination": ruleset1["rules"][0]["destination"],
-                "destination_type": ruleset1["rules"][0]["destination_type"],
-                "exit_uuid": str(uuid4()),
-            }
-        )
-
-        flow.update(flow_json)
-
-        self.assertFalse(RuleSet.objects.filter(flow=flow, uuid=ruleset1["uuid"]).exists())
-        self.assertTrue(ActionSet.objects.filter(flow=flow, uuid=ruleset1["uuid"]).exists())
-
-    def test_update_with_actionset_to_ruleset_change(self):
-        flow = self.get_flow("favorites")
-
-        # so that we check clearing of prefetched nodes that might be deleted in .update()
-        flow = Flow.objects.prefetch_related("action_sets", "rule_sets").get(id=flow.id)
-
-        flow_json = flow.as_json()
-
-        # remove first actionset
-        actionset1 = flow_json["action_sets"][0]
-        flow_json["action_sets"] = flow_json["action_sets"][1:]
-
-        # create ruleset in its place with same uuid etc
-        flow_json["rule_sets"].append(
-            {
-                "uuid": actionset1["uuid"],
-                "ruleset_type": "wait_message",
-                "rules": [
-                    {
-                        "uuid": actionset1["exit_uuid"],
-                        "test": {"type": "true"},
-                        "category": {"eng": "All Responses"},
-                        "destination": actionset1["destination"],
-                        "destination_type": "R",
-                    }
-                ],
-                "x": actionset1["x"],
-                "y": actionset1["y"],
-            }
-        )
-
-        flow.update(flow_json)
-
-        self.assertTrue(RuleSet.objects.filter(flow=flow, uuid=actionset1["uuid"]).exists())
-        self.assertFalse(ActionSet.objects.filter(flow=flow, uuid=actionset1["uuid"]).exists())
-
-    def test_ensure_current_version(self):
-        flow_json = self.get_flow_json("favorites_v4")["definition"]
-        flow = Flow.objects.create(
-            name="Favorites",
-            org=self.org,
-            created_by=self.admin,
-            modified_by=self.admin,
-            saved_by=self.admin,
-            version_number="4",
-            flow_type="M",
-        )
-
-        FlowRevision.objects.create(
-            flow=flow, definition=flow_json, spec_version=3, revision=1, created_by=self.admin, modified_by=self.admin
-        )
-
-        # now make sure we are on the latest version
-        flow.ensure_current_version()
-
-        # and that the format looks correct
-        flow_json = flow.as_json()
-
-        self.assertEqual(flow_json["metadata"]["name"], "Favorites")
-        self.assertEqual(flow_json["metadata"]["revision"], 2)
-        self.assertEqual(flow_json["metadata"]["expires"], 720)
-        self.assertEqual(flow_json["base_language"], "base")
-        self.assertEqual(len(flow_json["action_sets"]), 6)
-        self.assertEqual(len(flow_json["rule_sets"]), 6)
-
-    def test_migrate_to_11_12(self):
-        flow = self.get_flow("favorites")
-        definition = {
-            "entry": "79b4776b-a995-475d-ae06-1cab9af8a28e",
-            "rule_sets": [],
-            "action_sets": [
-                {
-                    "uuid": "d1244cfb-dc48-4dd5-ac45-7da49fdf46fb",
-                    "x": 459,
-                    "y": 150,
-                    "destination": "ef4865e9-1d34-4876-a0ff-fa3fe5025b3e",
-                    "actions": [
-                        {
-                            "type": "reply",
-                            "uuid": "3db54617-cce1-455b-a787-12df13df87bd",
-                            "msg": {"base": "Hi there"},
-                            "media": {},
-                            "quick_replies": [],
-                            "send_all": False,
-                        }
-                    ],
-                    "exit_uuid": "959fbe68-ba5a-4c78-b8d1-861e64d1e1e3",
-                },
-                {
-                    "uuid": "79b4776b-a995-475d-ae06-1cab9af8a28e",
-                    "x": 476,
-                    "y": 0,
-                    "destination": "d1244cfb-dc48-4dd5-ac45-7da49fdf46fb",
-                    "actions": [
-                        {
-                            "type": "channel",
-                            "uuid": "f133934a-9772-419f-ad52-00fe934dab19",
-                            "channel": None,
-                            "name": None,
-                        }
-                    ],
-                    "exit_uuid": "aec0318d-45c2-4c39-92fc-81d3d21178f6",
-                },
-            ],
-        }
-
-        migrated = migrate_to_version_11_12(definition, flow)
-
-        # removed the invalid reference
-        self.assertEqual(len(migrated["action_sets"]), 1)
-
-        # reconnected the nodes to new destinations and adjust entry
-        self.assertEqual(migrated["entry"], migrated["action_sets"][0]["uuid"])
-        self.assertEqual(migrated["action_sets"][0]["y"], 0)
-
-        definition = {
-            "entry": "79b4776b-a995-475d-ae06-1cab9af8a28e",
-            "rule_sets": [],
-            "action_sets": [
-                {
-                    "uuid": "d1244cfb-dc48-4dd5-ac45-7da49fdf46fb",
-                    "x": 459,
-                    "y": 150,
-                    "destination": "ef4865e9-1d34-4876-a0ff-fa3fe5025b3e",
-                    "actions": [
-                        {
-                            "type": "reply",
-                            "uuid": "3db54617-cce1-455b-a787-12df13df87bd",
-                            "msg": {"base": "Hi there"},
-                            "media": {},
-                            "quick_replies": [],
-                            "send_all": False,
-                        }
-                    ],
-                    "exit_uuid": "959fbe68-ba5a-4c78-b8d1-861e64d1e1e3",
-                },
-                {
-                    "uuid": "79b4776b-a995-475d-ae06-1cab9af8a28e",
-                    "x": 476,
-                    "y": 0,
-                    "destination": "d1244cfb-dc48-4dd5-ac45-7da49fdf46fb",
-                    "actions": [
-                        {
-                            "type": "channel",
-                            "uuid": "f133934a-9772-419f-ad52-00fe934dab19",
-                            "channel": self.channel.uuid,
-                            "name": self.channel.name,
-                        }
-                    ],
-                    "exit_uuid": "aec0318d-45c2-4c39-92fc-81d3d21178f6",
-                },
-            ],
-        }
-
-        migrated = migrate_to_version_11_12(definition, flow)
-
-        # removed the invalid reference
-        self.assertEqual(len(migrated["action_sets"]), 2)
-
-        flow = self.get_flow("migrate_to_11_12")
-        flow_json = self.get_flow_json("migrate_to_11_12")
-        migrate_to_version_11_12(flow_json, flow)
-
-        actionset = flow.action_sets.filter(y=0).first()
-        self.assertEqual(actionset.actions[0]["msg"]["base"], "Hey there, Yes or No?")
-
-        action_sets = flow.action_sets.all()
-        self.assertEqual(len(action_sets), 3)
-
-    def test_migrate_to_11_12_with_valid_channels(self):
-        self.channel.name = "1234"
-        self.channel.save()
-
-        self.org = self.channel.org
-        flow = self.get_flow("migrate_to_11_12")
-        flow_json = self.get_flow_json("migrate_to_11_12")
-        migrate_to_version_11_12(flow_json, flow)
-
-        action_sets = flow.action_sets.all()
-        self.assertEqual(len(action_sets), 7)
-
-    def test_migrate_to_11_12_with_one_node(self):
-
-        flow = self.get_flow("migrate_to_11_12_one_node")
-        flow_json = self.get_flow_json("migrate_to_11_12_one_node")
-        migrate_to_version_11_12(flow_json, flow)
-
-        action_sets = flow.action_sets.all()
-        self.assertEqual(len(action_sets), 0)
-
-    def test_migrate_to_11_12_other_org_new_flow(self):
-        # change ownership of the channel it's referencing
-        self.create_secondary_org()
-        self.channel.org = self.org2
-        self.channel.save(update_fields=("org",))
-
-        flow = self.get_flow("migrate_to_11_12_other_org", {"CHANNEL-UUID": str(self.channel.uuid)})
-
-        # check action was removed
-        definition = flow.revisions.order_by("revision").last().definition
-        self.assertEqual(len(definition["action_sets"]), 1)
-        self.assertEqual(len(definition["action_sets"][0]["actions"]), 0)
-
-    def test_migrate_to_11_12_other_org_existing_flow(self):
-        # import a flow but don't yet migrate it to 11.12
-        with patch("temba.flows.models.Flow.FINAL_LEGACY_VERSION", "11.11"):
-            flow = self.get_flow("migrate_to_11_12_other_org", {"CHANNEL-UUID": str(self.channel.uuid)})
-
-        self.assertEqual(flow.version_number, "11.11")
-        self.assertEqual(flow.revisions.order_by("revision").last().spec_version, "11.11")
-
-        # change ownership of the channel it's referencing
-        self.create_secondary_org()
-        self.channel.org = self.org2
-        self.channel.save(update_fields=("org",))
-
-        flow.ensure_current_version()
-
-        # check action set was removed
-        definition = flow.revisions.order_by("revision").last().definition
-        self.assertEqual(len(definition["action_sets"]), 0)
-
-    def test_migrate_to_11_11(self):
-
-        flow = self.get_flow("migrate_to_11_11")
-        flow_json = self.get_flow_json("migrate_to_11_11")
-
-        migrated = migrate_to_version_11_11(flow_json, flow)
-        migrated_labels = get_labels(migrated)
-        for uuid, name in migrated_labels.items():
-            self.assertTrue(Label.label_objects.filter(uuid=uuid, name=name).exists(), msg="Label UUID mismatch")
-
-    def test_migrate_to_11_10(self):
-        self.get_flow("migrate_to_11_10")
-
-        parent = Flow.objects.get(name__contains="Parent")
-        parent_json = parent.as_json()
-        ivr_child = Flow.objects.get(name__contains="IVR")
-
-        # the subflow ruleset to a messaging flow remains as the only ruleset
-        self.assertEqual(len(parent_json["rule_sets"]), 1)
-        self.assertEqual(parent_json["rule_sets"][0]["config"]["flow"]["name"], "Migrate to 11.10 SMS Child")
-
-        # whereas the subflow ruleset to an IVR flow has become a new trigger flow action
-        self.assertEqual(len(parent_json["action_sets"]), 4)
-
-        new_actionset = parent_json["action_sets"][3]
-        self.assertEqual(
-            new_actionset,
-            {
-                "uuid": matchers.UUID4String(),
-                "x": 218,
-                "y": 228,
-                "destination": parent_json["action_sets"][1]["uuid"],
-                "actions": [
-                    {
-                        "uuid": matchers.UUID4String(),
-                        "type": "trigger-flow",
-                        "flow": {"uuid": str(ivr_child.uuid), "name": "Migrate to 11.10 IVR Child"},
-                        "variables": [{"id": "@contact.uuid"}],
-                        "contacts": [],
-                        "groups": [],
-                    }
-                ],
-                "exit_uuid": matchers.UUID4String(),
-            },
-        )
-
-        # as did the start flow action
-        new_trigger2 = parent_json["action_sets"][1]["actions"][0]
-        self.assertEqual(
-            new_trigger2,
-            {
-                "uuid": matchers.UUID4String(),
-                "type": "trigger-flow",
-                "flow": {"uuid": str(ivr_child.uuid), "name": "Migrate to 11.10 IVR Child"},
-                "variables": [{"id": "@contact.uuid"}],
-                "contacts": [],
-                "groups": [],
-            },
-        )
-
-    def test_migrate_to_11_9(self):
-        self.get_flow("migrate_to_11_9")
-
-        invalid1 = Flow.objects.get(name="Invalid1")
-        invalid1.is_archived = True
-        invalid1.save()
-
-        invalid2 = Flow.objects.get(name="Invalid2")
-        invalid2.is_active = False
-        invalid2.save()
-
-        flow = Flow.objects.get(name="Master")
-        flow_json = flow.as_json()
-
-        self.assertEqual(len(flow_json["rule_sets"]), 4)
-        self.assertEqual(sum(len(action_set["actions"]) for action_set in flow_json["action_sets"]), 8)
-
-        migrated = migrate_to_version_11_9(flow_json, flow)
-
-        # expected to remove 1 ruleset and 3 actions referencing invalid flows
-        self.assertEqual(len(migrated["rule_sets"]), 3)
-        self.assertEqual(sum(len(action_set["actions"]) for action_set in migrated["action_sets"]), 5)
-
-    def test_migrate_to_11_8(self):
-        def get_rule_uuids(f):
-            uuids = []
-            for rs in f.get(Flow.RULE_SETS, []):
-                for rule in rs.get("rules"):
-                    uuids.append(rule["uuid"])
-            return uuids
-
-        original = self.get_flow_json("migrate_to_11_8")
-        original_uuids = get_rule_uuids(original)
-
-        self.assertEqual(len(original_uuids), 9)
-        self.assertEqual(len(set(original_uuids)), 7)
-
-        migrated = migrate_to_version_11_8(original)
-        migrated_uuids = get_rule_uuids(migrated)
-
-        # check that all rule UUIDs are now unique and only two new ones were added
-        self.assertEqual(len(set(migrated_uuids)), 9)
-        self.assertEqual(len(set(migrated_uuids).difference(original_uuids)), 2)
-
-    @override_settings(SEND_WEBHOOKS=True)
-    @uses_legacy_engine
-    def test_migrate_to_11_7(self):
-        original = self.get_flow_json("migrate_to_11_7")
-
-        self.assertEqual(len(original["action_sets"]), 5)
-        self.assertEqual(len(original["rule_sets"]), 1)
-
-        migrated = migrate_to_version_11_7(original)
-
-        self.assertEqual(len(migrated["action_sets"]), 3)
-        self.assertEqual(len(migrated["rule_sets"]), 6)
-
-        # run the flow to make sure the nodes are connected correctly
-        flow = self.get_flow("migrate_to_11_7")
-        with patch("requests.Session.send") as mock_request:
-            mock_request.return_value = MockResponse(200, "success")
-
-            run, = flow.start([], [self.contact])
-            Msg.create_incoming(self.channel, self.contact.get_urn().urn, "Ok")
-
-        self.assertEqual(flow.action_sets.count(), 3)
-
-        # check webhook calls made in correct order
-        self.assertEqual(len(mock_request.mock_calls), 5)
-        self.assertEqual(mock_request.mock_calls[0][1][0].url, "http://example.com/hook1")
-        self.assertEqual(mock_request.mock_calls[0][1][0].headers, {"Header1": "Value1", "User-agent": "RapidPro"})
-        self.assertEqual(mock_request.mock_calls[1][1][0].url, "http://example.com/hook2")
-        self.assertEqual(mock_request.mock_calls[2][1][0].url, "http://example.com/hook3")
-        self.assertEqual(mock_request.mock_calls[3][1][0].url, "http://example.com/hook4")
-        self.assertEqual(mock_request.mock_calls[4][1][0].url, "http://example.com/hook5")
-
-        # all migrated actions create a unique result value
-        run.refresh_from_db()
-        self.assertEqual(
-            set(run.results.keys()),
-            {
-                "response_1",
-                "migrated_webhook_1",
-                "migrated_webhook_2",
-                "migrated_webhook_3",
-                "migrated_webhook_4",
-                "migrated_webhook_5",
-            },
-        )
-        self.assertEqual(
-            run.results["migrated_webhook_5"],
-            {
-                "category": "Success",
-                "created_on": matchers.ISODate(),
-                "input": "GET http://example.com/hook5",
-                "name": "Migrated Webhook 5",
-                "node_uuid": matchers.UUID4String(),
-                "value": "200",
-            },
-        )
-
-    def test_migrate_to_11_6(self):
-
-        flow = self.get_flow("migrate_to_11_6")
-        flow_json = self.get_flow_json("migrate_to_11_6")
-
-        migrated = migrate_to_version_11_6(flow_json, flow)
-        migrated_groups = get_legacy_groups(migrated)
-        for uuid, name in migrated_groups.items():
-            self.assertTrue(ContactGroup.user_groups.filter(uuid=uuid, name=name).exists(), msg="Group UUID mismatch")
-
-    def test_migrate_to_11_5(self):
-        flow_json = self.get_flow_json("migrate_to_11_5")
-        flow_json = migrate_to_version_11_5(flow_json)
-
-        # check text was updated in the reply action
-        expected_msg = "\n".join(
-            [
-                "@extra.response_1",
-                "@extra.response_1",
-                "@flow.response_1.category",  # unchanged because its category
-                "@(upper(extra.response_1))",
-                "@(upper(flow.response_1.category))",
-                "",
-                "@flow.response_2",  # unchanged because this slug is also used by a non-webhook ruleset
-                "@flow.response_2.value",
-                "@flow.response_2.category",
-                "@(upper(flow.response_2))",
-                "@(upper(flow.response_2.category))",
-                "",
-                "@extra.response_3",
-                "@extra.response_3",
-                "@flow.response_3.category",
-                "@(upper(extra.response_3))",
-                "@(upper(flow.response_3.category))",
-            ]
-        )
-        self.assertEqual(flow_json["action_sets"][0]["actions"][0]["msg"]["eng"], expected_msg)
-
-        # check operand was updated in the split by expression
-        self.assertEqual(
-            flow_json["rule_sets"][4]["operand"], "@(extra.response_1 & flow.response_2 & extra.response_3)"
-        )
-
-        # check operand and type were updated in the split by flow field
-        rs = flow_json["rule_sets"][5]
-        self.assertEqual(rs["operand"], "@extra.response_1")
-        self.assertEqual(rs["ruleset_type"], "expression")
-
-        # check rule test was updated
-        self.assertEqual(flow_json["rule_sets"][5]["rules"][1]["test"]["test"]["eng"], "@extra.response_1")
-
-        # check webhook URL on ruleset was updated
-        self.assertEqual(flow_json["rule_sets"][6]["config"]["webhook"], "http://example.com/?thing=@extra.response_1")
-
-        # check webhook field on webhook action was updsated
-        self.assertEqual(
-            flow_json["action_sets"][1]["actions"][0]["webhook"], "http://example.com/?thing=@extra.response_1&foo=bar"
-        )
-
-        # check value field on save action was updsated
-        self.assertEqual(flow_json["action_sets"][1]["actions"][1]["value"], "@extra.response_3")
-
-    def test_migrate_to_11_4(self):
-        flow = self.get_flow("migrate_to_11_4")
-        flow_json = flow.as_json()
-
-        # gather up replies to check expressions were migrated
-        replies = []
-        for action_set in flow_json["action_sets"]:
-            for action in action_set["actions"]:
-                if "msg" in action:
-                    if isinstance(action["msg"], str):
-                        replies.append(action["msg"])
-                    else:
-                        for text in sorted(action["msg"].values()):
-                            replies.append(text)
-
-        self.assertEqual(
-            replies,
-            ['@flow.response_1.text\n@step.value\n@step.value\n@flow.response_3\n@(CONCATENATE(step.value, "blerg"))']
-            * 3,
-        )
-
-    def test_migrate_to_11_2(self):
-        fre_definition = {
-            "base_language": "fre",
-            "action_sets": [
-                {
-                    "uuid": "9468bbce-0df6-4d86-ae14-f26525ddda1d",
-                    "destination": "cc904a60-9de1-4f0b-9b55-a42b4ea6c434",
-                    "actions": [
-                        {
-                            "msg": {
-                                "base": "What is your favorite color?",
-                                "eng": "What is your favorite color?",
-                                "fra": "Quelle est votre couleur préférée?",
-                            },
-                            "type": "reply",
-                            "uuid": "335eb13d-5167-48ba-90c6-eb116656247c",
-                        }
-                    ],
-                    "exit_uuid": "a9904153-c831-4b95-aa20-13f84fed0841",
-                    "y": 0,
-                    "x": 100,
-                }
-            ],
-        }
-
-        base_definition = {
-            "base_language": "base",
-            "action_sets": [
-                {
-                    "uuid": "9468bbce-0df6-4d86-ae14-f26525ddda1d",
-                    "destination": "cc904a60-9de1-4f0b-9b55-a42b4ea6c434",
-                    "actions": [
-                        {
-                            "msg": {
-                                "base": "What is your favorite color?",
-                                "eng": "What is your favorite color?",
-                                "fra": "Quelle est votre couleur préférée?",
-                            },
-                            "type": "reply",
-                            "uuid": "335eb13d-5167-48ba-90c6-eb116656247c",
-                        }
-                    ],
-                    "exit_uuid": "a9904153-c831-4b95-aa20-13f84fed0841",
-                    "y": 0,
-                    "x": 100,
-                }
-            ],
-        }
-
-        flow1 = Flow.objects.create(
-            name="base lang test",
-            org=self.org,
-            created_by=self.admin,
-            modified_by=self.admin,
-            saved_by=self.admin,
-            version_number=1,
-        )
-        flow2 = Flow.objects.create(
-            name="Base lang test",
-            org=self.org,
-            created_by=self.admin,
-            modified_by=self.admin,
-            saved_by=self.admin,
-            version_number=1,
-        )
-        FlowRevision.objects.create(
-            flow=flow1,
-            definition=fre_definition,
-            spec_version=1,
-            revision=1,
-            created_by=self.admin,
-            modified_by=self.admin,
-        )
-        FlowRevision.objects.create(
-            flow=flow2,
-            definition=fre_definition,
-            spec_version=1,
-            revision=1,
-            created_by=self.admin,
-            modified_by=self.admin,
-        )
-
-        new_definition = migrate_to_version_11_2(fre_definition, flow=flow1)
-
-        fre_lang_value = new_definition["base_language"]
-        self.assertEqual(fre_lang_value, "fra")
-
-        new_definition = migrate_to_version_11_2(base_definition, flow=flow2)
-
-        base_lang_value = new_definition["base_language"]
-        self.assertEqual(base_lang_value, "base")
-
-    def test_migrate_to_11_1(self):
-        definition = {
-            "base_language": "base",
-            "action_sets": [
-                {
-                    "uuid": "9468bbce-0df6-4d86-ae14-f26525ddda1d",
-                    "destination": "cc904a60-9de1-4f0b-9b55-a42b4ea6c434",
-                    "actions": [
-                        {
-                            "msg": {
-                                "base": "What is your favorite color?",
-                                "eng": "What is your favorite color?",
-                                "fre": "Quelle est votre couleur préférée?",
-                            },
-                            "type": "reply",
-                            "uuid": "335eb13d-5167-48ba-90c6-eb116656247c",
-                        }
-                    ],
-                    "exit_uuid": "a9904153-c831-4b95-aa20-13f84fed0841",
-                    "y": 0,
-                    "x": 100,
-                },
-                {
-                    "y": 1214,
-                    "x": 284,
-                    "destination": "498b1953-02f1-47dd-b9cb-1b51913e348f",
-                    "uuid": "9769918c-8ca4-4ec5-8b5b-bf94cc6746a9",
-                    "actions": [
-                        {
-                            "lang": "fre",
-                            "type": "lang",
-                            "name": "French",
-                            "uuid": "56a4bca5-b9e5-4d04-883c-ca65d7c4d538",
-                        }
-                    ],
-                },
-                {
-                    "uuid": "9468bbce-0df6-4d86-ae14-f26525ddda1d",
-                    "destination": "cc904a60-9de1-4f0b-9b55-a42b4ea6c434",
-                    "actions": [
-                        {
-                            "msg": {
-                                "base": "What is your favorite color?",
-                                "eng": "What is your favorite color?",
-                                "fre": "Quelle est votre couleur préférée?",
-                                "newl": "Bogus translation",
-                            },
-                            "type": "reply",
-                            "uuid": "335eb13d-5167-48ba-90c6-eb116656247c",
-                        }
-                    ],
-                    "exit_uuid": "a9904153-c831-4b95-aa20-13f84fed0841",
-                    "y": 0,
-                    "x": 100,
-                },
-            ],
-        }
-
-        flow = Flow.objects.create(
-            name="String group",
-            org=self.org,
-            created_by=self.admin,
-            modified_by=self.admin,
-            saved_by=self.admin,
-            version_number=1,
-        )
-
-        FlowRevision.objects.create(
-            flow=flow, definition=definition, spec_version=1, revision=1, created_by=self.admin, modified_by=self.admin
-        )
-
-        new_definition = migrate_to_version_11_1(definition, flow=flow)
-
-        lang_path = new_definition["action_sets"][0]["actions"][0]["msg"]
-
-        self.assertIn("fra", lang_path)
-        self.assertEqual(len(lang_path), 3)
-
-        lang_key_value = new_definition["action_sets"][1]["actions"][0]["lang"]
-
-        self.assertEqual(lang_key_value, "fra")
-
-        should_not_be_migrated_path = new_definition["action_sets"][2]["actions"][0]["msg"]
-        self.assertIn("fre", should_not_be_migrated_path)
-
-        # we cannot migrate flows to version 11 without flow object (languages depend on flow.org)
-        self.assertRaises(ValueError, migrate_to_version_11_1, definition)
-
-    def test_migrate_to_11_0(self):
-        self.create_field("nickname", "Nickname", Value.TYPE_TEXT)
-        self.create_field("district", "District", Value.TYPE_DISTRICT)
-        self.create_field("joined_on", "Joined On", Value.TYPE_DATETIME)
-
-        flow = self.get_flow("type_flow")
-        flow_json = flow.as_json()
-
-        # gather up replies to check expressions were migrated
-        replies = []
-        for action_set in flow_json["action_sets"]:
-            for action in action_set["actions"]:
-                if action["type"] == "reply":
-                    for text in sorted(action["msg"].values()):
-                        replies.append(text)
-
-        self.assertEqual(
-            replies,
-            [
-                "Hey @contact.nickname, you joined on @(format_date(contact.joined_on)) in @(format_location(contact.district)).",
-                "It's @(format_date(date)). The time is @(format_date(date.now)) on @date.today.",
-                "Send text",
-                "You said @flow.text at @(format_date(flow.text.time)). Send date",
-                "You said @(format_date(flow.date)) which was in category @flow.date.category Send number",
-                "You said @flow.number. Send state",
-                "You said @(format_location(flow.state)) which was in category @flow.state.category. Send district",
-                "You said @(format_location(flow.district)). Send ward",
-                "Tu as dit @(format_location(flow.ward))",  # flow var followed by end of input
-                "You said @(format_location(flow.ward)).",  # flow var followed by period then end of input
-            ],
-        )
-
-    def test_migrate_to_11_0_with_null_ruleset_label(self):
-        flow = self.get_flow("migrate_to_11_0")
-        definition = {
-            "rule_sets": [
-                {
-                    "uuid": "9ed4a233-c737-4f46-9b0a-de6e88134e14",
-                    "rules": [],
-                    "ruleset_type": "wait_message",
-                    "label": None,
-                    "operand": None,
-                    "finished_key": None,
-                    "y": 180,
-                    "x": 179,
-                }
-            ]
-        }
-
-        migrated = migrate_to_version_11_0(definition, flow)
-
-        self.assertEqual(migrated, definition)
-
-    def test_migrate_to_11_0_with_null_msg_text(self):
-        flow = self.get_flow("migrate_to_11_0")
-        definition = {
-            "action_sets": [
-                {
-                    "y": 0,
-                    "x": 100,
-                    "destination": "0ecf7914-05e0-4b71-8816-495d2c0921b5",
-                    "uuid": "a6676605-332a-4309-a8b8-79b33e73adcd",
-                    "actions": [{"type": "reply", "msg": {"base": None}}],
-                }
-            ]
-        }
-
-        migrated = migrate_to_version_11_0(definition, flow)
-        self.assertEqual(migrated, definition)
-
-    def test_migrate_to_11_0_with_broken_localization(self):
-        migrated = self.get_flow("migrate_to_11_0").as_json()
-
-        self.assertEqual(
-            migrated["action_sets"][0]["actions"][0]["msg"],
-            {"base": "@(format_date(date)) Something went wrong once. I shouldn't be a dict inside a dict."},
-        )
-
-    def test_migrate_to_10_4(self):
-        definition = {
-            "action_sets": [
-                {
-                    "y": 0,
-                    "x": 100,
-                    "destination": "0ecf7914-05e0-4b71-8816-495d2c0921b5",
-                    "uuid": "a6676605-332a-4309-a8b8-79b33e73adcd",
-                    "actions": [{"type": "reply", "msg": {"base": "What is your favorite color?"}}],
-                }
-            ]
-        }
-
-        definition = migrate_to_version_10_4(definition)
-
-        # make sure all of our action sets have an exit uuid and all of our actions have uuids set
-        for actionset in definition["action_sets"]:
-            self.assertIsNotNone(actionset["exit_uuid"])
-            for action in actionset["actions"]:
-                self.assertIsNotNone(action["uuid"])
-
-    def test_migrate_to_10_3(self):
-        favorites = self.get_flow("favorites")
-
-        # make sure all of our action sets have an exit uuid
-        for actionset in favorites.action_sets.all():
-            self.assertIsNotNone(actionset.exit_uuid)
-
-    def test_migrate_to_10_2(self):
-        flow_json = self.get_flow_json("single_message_bad_localization")
-        flow_json = migrate_to_version_10_2(flow_json)
-        self.assertEqual("Campaign Message 12", flow_json["action_sets"][0]["actions"][0]["msg"]["eng"])
-
-    def test_migrate_to_10_1(self):
-        favorites = self.get_flow("favorites")
-
-        # make sure all of our actions have uuids set
-        for actionset in favorites.action_sets.all():
-            for action in actionset.get_actions():
-                self.assertIsNotNone(action.uuid)
-
-        # since actions can generate their own uuids, lets make sure fetching from the databse yields the same uuids
-        exported = favorites.as_json()
-        flow = Flow.objects.filter(name="Favorites").first()
-        self.assertEqual(exported, flow.as_json())
-        self.assertEqual(flow.version_number, Flow.FINAL_LEGACY_VERSION)
-
-    @override_settings(SEND_WEBHOOKS=True)
-    @uses_legacy_engine
-    def test_migrate_to_10(self):
-        # this is really just testing our rewriting of webhook rulesets
-        webhook_flow = self.get_flow("dual_webhook")
-        self.assertNotEqual(webhook_flow.modified_on, webhook_flow.saved_on)
-
-        # get our definition out
-        flow_def = webhook_flow.as_json()
-
-        # make sure our rulesets no longer have 'webhook' or 'webhook_action'
-        for ruleset in flow_def["rule_sets"]:
-            self.assertNotIn("webhook", ruleset)
-            self.assertNotIn("webhook_action", ruleset)
-
-        self.mockRequest("POST", "/code", '{"code": "ABABUUDDLRS"}', content_type="application/json")
-
-        run, = webhook_flow.start([], [self.contact])
-
-        # assert the code we received was right
-        msg = Msg.objects.filter(direction="O", contact=self.contact).order_by("id").last()
-        self.assertEqual(msg.text, "Great, your code is ABABUUDDLRS. Enter your name")
-
-        self.mockRequest("GET", "/success", "Success")
-
-        self.send_message(webhook_flow, "Ryan Lewis", assert_reply=False)
-
-        # startover have our first webhook fail, check that routing still works with failure
-        flow_def["rule_sets"][0]["config"]["webhook"] = "http://localhost:49999/error"
-        webhook_flow.update(flow_def)
-
-        self.mockRequest("POST", "/error", "BOOM", status=400)
-
-        webhook_flow.start([], [self.contact], restart_participants=True)
-
-        # assert the code we received was right
-        msg = Msg.objects.filter(direction="O", contact=self.contact).order_by("id").last()
-        self.assertEqual("Great, your code is @extra.code. Enter your name", msg.text)
-
-        # check all our mocked requests were made
-        self.assertAllRequestsMade()
-
-    def test_migrate_to_9(self):
-
-        # our group and flow to move to uuids
-        group = self.create_group("Phans", [])
-        previous_flow = self.create_flow()
-        start_flow = self.create_flow()
-        label = Label.get_or_create(self.org, self.admin, "My label")
-
-        substitutions = dict(
-            group_id=group.pk,
-            contact_id=self.contact.pk,
-            start_flow_id=start_flow.pk,
-            previous_flow_id=previous_flow.pk,
-            label_id=label.pk,
-        )
-
-        exported_json = self.get_import_json("migrate_to_9", substitutions)
-        exported_json = migrate_export_to_version_9(exported_json, self.org, True)
-
-        # our campaign events shouldn't have ids
-        campaign = exported_json["campaigns"][0]
-        event = campaign["events"][0]
-
-        # campaigns should have uuids
-        self.assertIn("uuid", campaign)
-        self.assertNotIn("id", campaign)
-
-        # our event flow should be a uuid
-        self.assertIn("flow", event)
-        self.assertIn("uuid", event["flow"])
-        self.assertNotIn("id", event["flow"])
-
-        # our relative field should not have an id
-        self.assertNotIn("id", event["relative_to"])
-
-        # evaluate that the flow json is migrated properly
-        flow_json = exported_json["flows"][0]
-
-        # check that contacts migrated properly
-        send_action = flow_json["action_sets"][0]["actions"][1]
-        self.assertEqual(1, len(send_action["contacts"]))
-        self.assertEqual(0, len(send_action["groups"]))
-
-        for contact in send_action["contacts"]:
-            self.assertIn("uuid", contact)
-            self.assertNotIn("id", contact)
-
-        for group in send_action["groups"]:
-            if isinstance(group, dict):
-                self.assertIn("uuid", group)
-                self.assertNotIn("id", group)
-
-        label_action = flow_json["action_sets"][0]["actions"][2]
-        for label in label_action.get("labels"):
-            self.assertNotIn("id", label)
-            self.assertIn("uuid", label)
-
-        action_set = flow_json["action_sets"][1]
-        actions = action_set["actions"]
-
-        for action in actions[0:2]:
-            self.assertIn(action["type"], ("del_group", "add_group"))
-            self.assertIn("uuid", action["groups"][0])
-            self.assertNotIn("id", action["groups"][0])
-
-        for action in actions[2:4]:
-            self.assertIn(action["type"], ("trigger-flow", "flow"))
-            self.assertIn("flow", action)
-            self.assertIn("uuid", action["flow"])
-            self.assertIn("name", action["flow"])
-            self.assertNotIn("id", action)
-            self.assertNotIn("name", action)
-
-        # we also switch flow ids to uuids in the metadata
-        self.assertIn("uuid", flow_json["metadata"])
-        self.assertNotIn("id", flow_json["metadata"])
-
-        # import the same thing again, should have the same uuids
-        new_exported_json = self.get_import_json("migrate_to_9", substitutions)
-        new_exported_json = migrate_export_to_version_9(new_exported_json, self.org, True)
-        self.assertEqual(flow_json["metadata"]["uuid"], new_exported_json["flows"][0]["metadata"]["uuid"])
-
-        # but when done as a different site, it should be unique
-        new_exported_json = self.get_import_json("migrate_to_9", substitutions)
-        new_exported_json = migrate_export_to_version_9(new_exported_json, self.org, False)
-        self.assertNotEqual(flow_json["metadata"]["uuid"], new_exported_json["flows"][0]["metadata"]["uuid"])
-
-        # check we can update a flow with the migrated definition
-        flow = Flow.objects.create(
-            name="test flow", created_by=self.admin, modified_by=self.admin, org=self.org, saved_by=self.admin
-        )
-        flow.update(FlowRevision.migrate_definition(exported_json["flows"][0], flow))
-
-        # can also just import a single flow
-        exported_json = self.get_import_json("migrate_to_9", substitutions)
-        flow_json = migrate_to_version_9(exported_json["flows"][0], flow)
-        self.assertIn("uuid", flow_json["metadata"])
-        self.assertNotIn("id", flow_json["metadata"])
-
-        # try it with missing metadata
-        flow_json = self.get_import_json("migrate_to_9", substitutions)["flows"][0]
-        del flow_json["metadata"]
-        flow_json = migrate_to_version_9(flow_json, flow)
-        self.assertEqual(1, flow_json["metadata"]["revision"])
-        self.assertEqual("test flow", flow_json["metadata"]["name"])
-        self.assertEqual(720, flow_json["metadata"]["expires"])
-        self.assertIn("uuid", flow_json["metadata"])
-        self.assertIn("saved_on", flow_json["metadata"])
-
-        # check that our replacements work
-        self.assertEqual("@(CONCAT(parent.divided, parent.sky))", flow_json["action_sets"][0]["actions"][3]["value"])
-        self.assertEqual("@parent.contact.name", flow_json["action_sets"][0]["actions"][4]["value"])
-
-    def test_migrate_to_8(self):
-        # file uses old style expressions
-        flow_json = self.get_flow_json("old_expressions")
-
-        # migrate to the version right before us first
-        flow_json = migrate_to_version_7(flow_json)
-        flow_json = migrate_to_version_8(flow_json)
-
-        self.assertEqual(
-            flow_json["action_sets"][0]["actions"][0]["msg"]["eng"], "Hi @(UPPER(contact.name)). Today is @(date.now)"
-        )
-        self.assertEqual(flow_json["action_sets"][1]["actions"][0]["groups"][0], "@flow.response_1.category")
-        self.assertEqual(
-            flow_json["action_sets"][1]["actions"][1]["msg"]["eng"], "Was @(PROPER(LOWER(contact.name)))."
-        )
-        self.assertEqual(flow_json["action_sets"][1]["actions"][1]["variables"][0]["id"], "@flow.response_1.category")
-        self.assertEqual(
-            flow_json["rule_sets"][0]["webhook"], "http://example.com/query.php?contact=@(UPPER(contact.name))"
-        )
-        self.assertEqual(flow_json["rule_sets"][0]["operand"], "@(step.value)")
-        self.assertEqual(flow_json["rule_sets"][1]["operand"], "@(step.value + 3)")
-
-    def test_migrate_to_7(self):
-        flow_json = self.get_flow_json("ivr_v3")
-
-        # migrate to the version right before us first
-        flow_json = migrate_to_version_5(flow_json)
-        flow_json = migrate_to_version_6(flow_json)
-
-        self.assertIsNotNone(flow_json.get("definition"))
-        self.assertEqual("Call me maybe", flow_json.get("name"))
-        self.assertEqual(100, flow_json.get("id"))
-        self.assertEqual("V", flow_json.get("flow_type"))
-
-        flow_json = migrate_to_version_7(flow_json)
-        self.assertIsNone(flow_json.get("definition", None))
-        self.assertIsNotNone(flow_json.get("metadata", None))
-
-        metadata = flow_json.get("metadata")
-        self.assertEqual("Call me maybe", metadata["name"])
-        self.assertEqual(100, metadata["id"])
-        self.assertEqual("V", flow_json.get("flow_type"))
-
-    def test_migrate_to_6(self):
-
-        # file format is old non-localized format
-        voice_json = self.get_flow_json("ivr_v3")
-        definition = voice_json.get("definition")
-
-        # no language set
-        self.assertIsNone(definition.get("base_language", None))
-        self.assertEqual("Yes", definition["rule_sets"][0]["rules"][0]["category"])
-        self.assertEqual("Press one, two, or three. Thanks.", definition["action_sets"][0]["actions"][0]["msg"])
-
-        # add a recording to make sure that gets migrated properly too
-        definition["action_sets"][0]["actions"][0]["recording"] = "/recording.mp3"
-
-        voice_json = migrate_to_version_5(voice_json)
-        voice_json = migrate_to_version_6(voice_json)
-        definition = voice_json.get("definition")
-
-        # now we should have a language
-        self.assertEqual("base", definition.get("base_language", None))
-        self.assertEqual("Yes", definition["rule_sets"][0]["rules"][0]["category"]["base"])
-        self.assertEqual(
-            "Press one, two, or three. Thanks.", definition["action_sets"][0]["actions"][0]["msg"]["base"]
-        )
-        self.assertEqual("/recording.mp3", definition["action_sets"][0]["actions"][0]["recording"]["base"])
-
-        # now try one that doesn't have a recording set
-        voice_json = self.get_flow_json("ivr_v3")
-        definition = voice_json.get("definition")
-        del definition["action_sets"][0]["actions"][0]["recording"]
-        voice_json = migrate_to_version_5(voice_json)
-        voice_json = migrate_to_version_6(voice_json)
-        definition = voice_json.get("definition")
-        self.assertNotIn("recording", definition["action_sets"][0]["actions"][0])
-
-    def test_migrate_to_5_language(self):
-
-        flow_json = self.get_flow_json("multi_language_flow")
-        ruleset = flow_json["definition"]["rule_sets"][0]
-        ruleset["operand"] = "@step.value|lower_case"
-
-        # now migrate us forward
-        flow_json = migrate_to_version_5(flow_json)
-
-        wait_ruleset = None
-        rules = None
-        for ruleset in flow_json.get("definition").get("rule_sets"):
-            if ruleset["ruleset_type"] == "wait_message":
-                rules = ruleset["rules"]
-                wait_ruleset = ruleset
-                break
-
-        self.assertIsNotNone(wait_ruleset)
-        self.assertIsNotNone(rules)
-
-        self.assertEqual(1, len(rules))
-        self.assertEqual("All Responses", rules[0]["category"]["eng"])
-        self.assertEqual("Otro", rules[0]["category"]["spa"])
-
-    @override_settings(SEND_WEBHOOKS=True)
-    @uses_legacy_engine
-    def test_migrate_to_5(self):
-        flow = self.get_flow("favorites_v4")
-
-        # first node should be a wait node
-        ruleset = RuleSet.objects.filter(label="Color Response").first()
-        self.assertEqual("wait_message", ruleset.ruleset_type)
-        self.assertEqual("@step.value", ruleset.operand)
-
-        # we should now be pointing to a newly created webhook rule
-        webhook = RuleSet.objects.get(flow=flow, uuid=ruleset.get_rules()[0].destination)
-        self.assertEqual("webhook", webhook.ruleset_type)
-        self.assertEqual("http://localhost:49999/status", webhook.config[RuleSet.CONFIG_WEBHOOK])
-        self.assertEqual("POST", webhook.config[RuleSet.CONFIG_WEBHOOK_ACTION])
-        self.assertEqual("@step.value", webhook.operand)
-        self.assertEqual("Color Webhook", webhook.label)
-
-        # which should in turn point to a new expression split on @extra.value
-        expression = RuleSet.objects.get(flow=flow, uuid=webhook.get_rules()[0].destination)
-        self.assertEqual("expression", expression.ruleset_type)
-        self.assertEqual("@extra.value", expression.operand)
-
-        # takes us to the next question
-        beer_question = ActionSet.objects.get(flow=flow, uuid=expression.get_rules()[0].destination)
-
-        # which should pause for the response
-        wait_beer = RuleSet.objects.get(flow=flow, uuid=beer_question.destination)
-        self.assertEqual("wait_message", wait_beer.ruleset_type)
-        self.assertEqual("@step.value", wait_beer.operand)
-        self.assertEqual(1, len(wait_beer.get_rules()))
-        self.assertEqual("All Responses", wait_beer.get_rules()[0].category[flow.base_language])
-
-        # and then split on the expression for various beer choices
-        beer_expression = RuleSet.objects.get(flow=flow, uuid=wait_beer.get_rules()[0].destination)
-        self.assertEqual("expression", beer_expression.ruleset_type)
-        self.assertEqual("@(LOWER(step.value))", beer_expression.operand)
-        self.assertEqual(5, len(beer_expression.get_rules()))
-
-        # set our expression to operate on the last inbound message
-        expression.operand = "@step.value"
-        expression.save()
-
-        # now try executing our migrated flow
-        first_response = ActionSet.objects.get(flow=flow, x=131)
-        actions = first_response.actions
-        actions[0]["msg"][
-            flow.base_language
-        ] = "I like @flow.color.category too! What is your favorite beer? @flow.color_webhook"
-        first_response.actions = actions
-        first_response.save()
-
-        self.mockRequest("POST", "/status", '{ "status": "valid" }')
-
-        reply = self.send_message(flow, "red")
-        self.assertEqual("I like Red too! What is your favorite beer? 200", reply)
-
-        reply = self.send_message(flow, "Turbo King")
-        self.assertEqual(
-            "Mmmmm... delicious Turbo King. If only they made red Turbo King! Lastly, what is your name?", reply
-        )
-
-        # check all our mocked requests were made
-        self.assertAllRequestsMade()
-
-    def test_migrate_revisions(self):
-        flow = self.get_flow("favorites_v4")
-        rev = flow.revisions.all().first()
-        json_flow = rev.get_definition_json()
-
-        # remove our flow version from the flow
-        del json_flow[Flow.VERSION]
-        rev.definition = json_flow
-        rev.spec_version = "10"
-        rev.save()
-
-        new_rev = flow.update(rev.get_definition_json())
-        self.assertEqual(new_rev.spec_version, Flow.FINAL_LEGACY_VERSION)
-
-        flow.refresh_from_db()
-        self.assertEqual(flow.revisions.all().count(), 2)
-        self.assertEqual(flow.version_number, Flow.FINAL_LEGACY_VERSION)
-
-    def test_migrate_sample_flows(self):
-        self.org.create_sample_flows("https://app.rapidpro.io")
-        self.assertEqual(3, self.org.flows.filter(name__icontains="Sample Flow").count())
-
-        # make sure it is localized
-        poll = self.org.flows.filter(name="Sample Flow - Simple Poll").first()
-        self.assertIn("base", poll.action_sets.all().order_by("y").first().get_actions()[0].msg)
-        self.assertEqual("base", poll.base_language)
-
-        # check replacement
-        order_checker = self.org.flows.filter(name="Sample Flow - Order Status Checker").first()
-        ruleset = order_checker.rule_sets.filter(y=298).first()
-        self.assertEqual("https://app.rapidpro.io/demo/status/", ruleset.config[RuleSet.CONFIG_WEBHOOK])
-
-        # our test user doesn't use an email address, check for Administrator for the email
-        actionset = order_checker.action_sets.filter(y=991).first()
-        self.assertEqual("Administrator", actionset.get_actions()[1].emails[0])
-
-    def test_migrate_bad_group_names(self):
-        # This test makes sure that bad contact groups (< 25, etc) are migrated forward properly.
-        # However, since it was a missed migration, now we need to apply it for any current version
-        # at the time of this fix
-        for v in ("4", "5", "6", "7", "8", "9", "10"):
-            error = 'Failure migrating group names "%s" forward from v%s'
-            flow = self.get_flow("favorites_bad_group_name_v%s" % v)
-            self.assertIsNotNone(flow, "Failure importing favorites from v%s" % v)
-            self.assertTrue(ContactGroup.user_groups.filter(name="Contacts < 25").exists(), error % ("< 25", v))
-            self.assertTrue(ContactGroup.user_groups.filter(name="Contacts > 100").exists(), error % ("> 100", v))
-
-            ContactGroup.user_groups.all().delete()
-            self.assertEqual(Flow.FINAL_LEGACY_VERSION, flow.version_number)
-            flow.release()
-
-    def test_migrate_malformed_groups(self):
-        flow = self.get_flow("malformed_groups")
-        self.assertIsNotNone(flow)
-        self.assertTrue(ContactGroup.user_groups.filter(name="Contacts < 25").exists())
-        self.assertTrue(ContactGroup.user_groups.filter(name="Unknown").exists())
-
-
 class DuplicateResultTest(FlowFileTest):
     @uses_legacy_engine
     def test_duplicate_value_test(self):
@@ -10271,7 +8930,7 @@ class ChannelSplitTest(FlowFileTest):
         flow = self.get_flow("channel_split")
 
         # start our contact down the flow
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         # check the message sent to them
         msgs = list(self.contact.msgs.order_by("id"))
@@ -10287,7 +8946,7 @@ class ChannelSplitTest(FlowFileTest):
         self.contact.urns.all().update(contact=None)
 
         # run the flow again
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         # shouldn't have any messages sent, as they have no URN
         self.assertFalse(self.contact.msgs.all())
@@ -10301,7 +8960,7 @@ class ChannelSplitTest(FlowFileTest):
         flow = self.get_flow("channel_split_rule_first")
 
         # start our contact down the flow
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         # check that the split was successful
         msg = self.contact.msgs.first()
@@ -10332,7 +8991,7 @@ class GhostActionNodeTest(FlowFileTest):
         flow = Flow.objects.get(name="Parent Flow")
 
         # start the flow
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         # at this point, our contact has to active flow runs:
         # one for our parent flow at an action set (the start flow action), one in our child flow at the send message action
@@ -10342,7 +9001,7 @@ class GhostActionNodeTest(FlowFileTest):
 
         # create a new message and get it handled
         msg = self.create_msg(contact=self.contact, direction="I", text="yes")
-        Flow.find_and_handle(msg)
+        legacy.find_and_handle(msg)
 
         # we should have gotten a response from our child flow
         self.assertEqual(
@@ -10354,7 +9013,7 @@ class TwoInRowTest(FlowFileTest):
     @uses_legacy_engine
     def test_two_in_row(self):
         flow = self.get_flow("two_in_row")
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         # assert contact received both messages
         msgs = self.contact.msgs.all()
@@ -10396,7 +9055,7 @@ class FlowSessionCRUDLTest(TembaTest):
     def test_session_json(self):
         contact = self.create_contact("Bob", number="+1234567890")
         flow = self.get_flow("color")
-        flow.start([], [contact])
+        legacy.flow_start(flow, [], [contact])
 
         # create a fake session for this run
         session = FlowSession.objects.create(
@@ -10432,14 +9091,14 @@ class ExitTest(FlowFileTest):
     def test_exit_via_start(self):
         # start contact in one flow
         first_flow = self.get_flow("substitution")
-        first_flow.start([], [self.contact])
+        legacy.flow_start(first_flow, [], [self.contact])
 
         # should have one active flow run
         first_run = FlowRun.objects.get(is_active=True, flow=first_flow, contact=self.contact)
 
         # start in second via manual start
         second_flow = self.get_flow("favorites")
-        second_flow.start([], [self.contact])
+        legacy.flow_start(second_flow, [], [self.contact])
 
         second_run = FlowRun.objects.get(is_active=True)
         first_run.refresh_from_db()
@@ -10449,90 +9108,20 @@ class ExitTest(FlowFileTest):
         self.assertTrue(second_run.is_active)
 
 
-class MigrationUtilsTest(TembaTest):
-    def test_map_actions(self):
-        # minimalist flow def with just actions and entry
-        flow_def = dict(
-            entry="1234",
-            action_sets=[dict(uuid="1234", x=100, y=0, actions=[dict(type="reply", msg=None)])],
-            rule_sets=[dict(y=10, x=100, uuid="5678")],
-        )
-        removed = map_actions(flow_def, lambda x: None)
-
-        # no more action sets and entry is remapped
-        self.assertFalse(removed["action_sets"])
-        self.assertEqual("5678", removed["entry"])
-
-        # add two action sets, we should remap entry to be the first
-        flow_def["action_sets"] = [
-            dict(uuid="1234", y=0, x=100, actions=[dict(type="reply", msg=None)]),
-            dict(uuid="2345", y=5, x=100, actions=[dict(type="reply", msg="foo")]),
-        ]
-        removed = map_actions(flow_def, lambda x: None if x["msg"] is None else x)
-
-        self.assertEqual(len(removed["action_sets"]), 1)
-        self.assertEqual(removed["action_sets"][0]["uuid"], "2345")
-        self.assertEqual(removed["entry"], "2345")
-
-        # remove a single action
-        flow_def["action_sets"] = [
-            dict(uuid="1234", y=10, x=100, actions=[dict(type="reply", msg=None), dict(type="reply", msg="foo")])
-        ]
-        removed = map_actions(flow_def, lambda x: None if x["msg"] is None else x)
-
-        self.assertEqual(len(removed["action_sets"]), 1)
-        self.assertEqual(len(removed["action_sets"][0]["actions"]), 1)
-        self.assertEqual(removed["entry"], "2345")
-
-        # no entry
-        flow_def = dict(
-            entry="1234",
-            action_sets=[dict(uuid="1234", y=0, x=100, actions=[dict(type="reply", msg=None)])],
-            rule_sets=[],
-        )
-        removed = map_actions(flow_def, lambda x: None if x["msg"] is None else x)
-
-        self.assertEqual(len(removed["action_sets"]), 0)
-        self.assertEqual(removed["entry"], None)
-
-        # check entry horizontal winner
-        flow_def = dict(
-            entry="1234",
-            action_sets=[dict(uuid="1234", x=100, y=0, actions=[dict(type="reply", msg=None)])],
-            rule_sets=[dict(y=10, x=100, uuid="5678"), dict(y=10, x=50, uuid="9012")],
-        )
-        removed = map_actions(flow_def, lambda x: None if x["msg"] is None else x)
-        self.assertEqual(removed["entry"], "9012")
-
-        # same horizontal check with action sets
-        flow_def = dict(
-            entry="1234",
-            action_sets=[
-                dict(uuid="1234", x=100, y=0, actions=[dict(type="reply", msg=None)]),
-                dict(uuid="9012", x=50, y=50, actions=[dict(type="reply", msg="foo")]),
-                dict(uuid="3456", x=0, y=50, actions=[dict(type="reply", msg="foo")]),
-            ],
-            rule_sets=[dict(y=100, x=100, uuid="5678")],
-        )
-
-        removed = map_actions(flow_def, lambda x: None if x["msg"] is None else x)
-        self.assertEqual(removed["entry"], "3456")
-
-
 class TriggerFlowTest(FlowFileTest):
     @uses_legacy_engine
     def test_trigger_then_loop(self):
         # start our parent flow
         flow = self.get_flow("parent_child_loop")
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         # trigger our second flow to start
         msg = self.create_msg(contact=self.contact, direction="I", text="add 12067797878")
-        Flow.find_and_handle(msg)
+        legacy.find_and_handle(msg)
 
         child_run = FlowRun.objects.get(contact__urns__path="+12067797878")
         msg = self.create_msg(contact=child_run.contact, direction="I", text="Christine")
-        Flow.find_and_handle(msg)
+        legacy.find_and_handle(msg)
         child_run.refresh_from_db()
         self.assertEqual("C", child_run.exit_type)
 
@@ -10543,7 +9132,7 @@ class TriggerFlowTest(FlowFileTest):
 
         # and can do it again
         msg = self.create_msg(contact=self.contact, direction="I", text="add 12067798080")
-        Flow.find_and_handle(msg)
+        legacy.find_and_handle(msg)
 
         FlowRun.objects.get(contact__urns__path="+12067798080")
         run.refresh_from_db()
@@ -10571,7 +9160,7 @@ class StackedExitsTest(FlowFileTest):
         self.get_flow("stacked_exits")
         flow = Flow.objects.get(name="Stacked")
 
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         msgs = Msg.objects.filter(contact=self.contact).order_by("sent_on")
         self.assertEqual(3, msgs.count())
@@ -10592,7 +9181,7 @@ class StackedExitsTest(FlowFileTest):
         self.get_flow("stacked_webhook_exits")
         flow = Flow.objects.get(name="Stacked")
 
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         msgs = Msg.objects.filter(contact=self.contact).order_by("sent_on")
         self.assertEqual(4, msgs.count())
@@ -10614,7 +9203,7 @@ class StackedExitsTest(FlowFileTest):
         self.get_flow("stacked_response_exits")
         flow = Flow.objects.get(name="Stacked")
 
-        flow.start([], [self.contact])
+        legacy.flow_start(flow, [], [self.contact])
 
         msgs = Msg.objects.filter(contact=self.contact).order_by("sent_on")
         self.assertEqual(2, msgs.count())
@@ -10683,10 +9272,12 @@ class TypeTest(TembaTest):
         self.assertEqual(Value.TYPE_WARD, RuleSet.objects.get(label="Ward").value_type)
 
         incoming = self.create_msg(direction=INCOMING, contact=contact, text="types")
-        flow.start(groups=[], contacts=[contact], start_msg=incoming)
+        legacy.flow_start(flow, groups=[], contacts=[contact], start_msg=incoming)
 
-        self.assertTrue(Flow.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="Some Text")))
-        self.assertTrue(Flow.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="not a date")))
+        self.assertTrue(legacy.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="Some Text")))
+        self.assertTrue(
+            legacy.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="not a date"))
+        )
 
         results = FlowRun.objects.get().results
 
@@ -10701,19 +9292,21 @@ class TypeTest(TembaTest):
         self.assertEqual("Other", results["date"]["category"])
 
         self.assertTrue(
-            Flow.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="Born 23/06/1977"))
+            legacy.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="Born 23/06/1977"))
         )
         self.assertTrue(
-            Flow.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="The number is 10"))
+            legacy.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="The number is 10"))
         )
         self.assertTrue(
-            Flow.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="I'm in Eastern Province"))
+            legacy.find_and_handle(
+                self.create_msg(contact=contact, direction=INCOMING, text="I'm in Eastern Province")
+            )
         )
         self.assertTrue(
-            Flow.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="That's in Gatsibo"))
+            legacy.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="That's in Gatsibo"))
         )
         self.assertTrue(
-            Flow.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="ya ok that's Kageyo"))
+            legacy.find_and_handle(self.create_msg(contact=contact, direction=INCOMING, text="ya ok that's Kageyo"))
         )
 
         results = FlowRun.objects.get().results
